\name{age.length.key}
\alias{age.length.key}
\title{Age length key}
\usage{age.length.key(sim, age.agg, len.agg)}
\description{Calculates the age-length-key for the survey and commercial fleet.}
\value{Dataframe containing the age - length key}
\author{Bjarki Þór Elvarsson}
\arguments{\item{sim}{Results from a Rgadget simulation}
\item{age.agg}{The desired age aggregation}
\item{len.agg}{The desired length aggregation}}
