% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{aldist}
\alias{aldist}
\title{Age length dist}
\usage{
aldist(stock.dat, sigma = 0, dl = 1)
}
\arguments{
\item{stock.dat}{}

\item{sigma}{}

\item{dl}{}
}
\value{
data.frame
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

