% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{callParamin}
\alias{callParamin}
\title{Call paramin}
\usage{
callParamin(i = "params.in", func = "gadget -s -n", opt = "optinfo",
  network = "network", o = "params.out", scale = NULL, condor = NULL,
  paramin.exe = "paramin")
}
\arguments{
\item{i}{}

\item{func}{}

\item{opt}{}

\item{network}{}

\item{o}{}

\item{scale}{}

\item{condor}{}

\item{paramin.exe}{}
}
\value{
null
}
\description{
Probably defunct
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

