% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catch.in.kilos}
\alias{catch.in.kilos}
\title{Catch in Kilos}
\usage{
catch.in.kilos(sim)
}
\arguments{
\item{sim}{Results from a Rgadget simulation}
}
\value{
Dataframe with the catch in kilos by timestep and ared.
}
\description{
Calculates the overall weigth of the catches by time step and area.
}
\author{
Bjarki Þór Elvarsson
}

