% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clear.spaces}
\alias{clear.spaces}
\title{Clear spaces}
\usage{
clear.spaces(text)
}
\arguments{
\item{text}{string}
}
\value{
list or matrix containing the (non-empty) values from the string
}
\description{
Clear tab and spaces from a string and return a list or a matrix of values
}
\author{
Bjarki ??r Elvarsson
}

