% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gadget.forward}
\alias{gadget.forward}
\alias{gadget.ypr}
\title{Deprecated functions}
\usage{
gadget.forward(...)

gadget.ypr(...)
}
\arguments{
\item{...}{unused inputs}
}
\description{
Deprecated functions
}
