\name{derivedOptions}
\alias{derivedOptions}
\title{Derived options}
\usage{derivedOptions(opt)}
\description{This function is a helper function for RGadget and all file export functions,
it calculates all additional options and switches that can be derived from
the gadget options lists.}
\value{augmented gadget options list}
\arguments{\item{opt}{gadget options list}}
