% Generated by roxygen2 (4.0.2): do not edit by hand
\name{distr}
\alias{distr}
\title{Length distribution}
\usage{
distr(mu, sigma, l)
}
\arguments{
\item{mu}{mean length for all ages}

\item{sigma}{standard deviation of length for all ages}

\item{l}{lengthgroups}
}
\value{
a matrix of dimension length(mu) X (length(l)-1)
}
\description{
This is a helper function for the firststep function. This defines the
length distribution for each age group
}

