% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{eat}
\alias{eat}
\title{Eat}
\usage{
eat(PreyInArea, PredInArea, step, opt)
}
\arguments{
\item{PreyInArea}{Number of prey items in the area}

\item{PredInArea}{Number of predator items in the area}

\item{step}{the timestep, that is the time of the year}

\item{opt}{gadget options list}
}
\value{
The total unadjusted consumption of the predator on the prey
}
\description{
The following variables are used in the consumption calculations:
l is Lengthgroup in prey
L is Lengthgroup in predator
A is Areasize
H The density (biomass per area unit) of available food at
which the predator can consume half maximum consumption
\eqn{\Delta t} as Length of timestep
\eqn{M_{pred}(L)} as Maximum consumption
\eqn{\psi_{pred}(L)} as Fraction of \eqn{M_{pred}} consumed
N_{pred}(L) as Number of predator \eqn{pred} in lengthgroup $L
N_{prey}(l) as Number of prey \eqn{prey} in lengthgroup l
W_{prey}(l) as The mean weight of prey of length l
S_{pred,prey}(L,l) as Suitability of prey at length l for pred at length L
C_{prey,pred}(L,L) as Total weight predator of length L consumes of prey of length $l$
The formula for the consumption is as follows:
\deqn{C_{pred,prey}(L,l)=N_{pred}(L)M_{pred}(L)\Psi_{pred}(L)\frac{F_{pred,prey}(L,l)}{\sum_{l,prey}F_{pred,prey}(L,l)}}
\deqn{=N_{pred}(L)M_{pred}(L)\frac{\sum_{l,prey}F_{pred,prey}(L,l)}{\sum_{l,prey}F_{pred,prey}(L,l)+HA}\frac{F_{pred,prey}(L,l)}{\sum_{l,prey}F_{pred,prey}(L,l)}}
\deqn{=N_{pred}(L)M_{pred}(L)\frac{F_{pred,prey}(L,l)}{\sum_{l,prey}F_{pred,prey}(L,l)+HA}}
where
\deqn{F_{pred,prey}(L,l) =S_{pred,prey}(L,l)N_{prey}(l)W_{prey}(l)}
\deqn{ M_{pred}(L) =m_0e^{(m_1T-m_2T^3)}L_{pred}^{m_4}\Delta t}
The suitability function for predation used in the \R model is:
\deqn{S_{pred,prey}(L,l) = \frac{\delta}{1+e^{-\alpha-\beta l-\gamma L}}}
With one predator, one prey and otherfood the equation becomes:
\deqn{C_{L,l}=N_{L}M_{L}\Psi_{L}\frac{F_{L,l}}{\sum_lF_{L,l}+OA}}
\deqn{=N_{L}M_{L}\frac{F_{L,l}}{\sum_lF_{L,l}+OA+HA}}
where O is the density of otherfood.
}
