% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eval.gadget.formula}
\alias{eval.gadget.formula}
\title{eval.gadget.formula}
\usage{
eval.gadget.formula(gad.for, par)
}
\arguments{
\item{gad.for}{gadget formula}

\item{par}{gadget parameters object}
}
\value{
a vector of evaluated gadget formulas
}
\description{
Evaluate gadget formulas, which are in reverse polish notation, ie
'(* x y)' which is equivalent to 'x*y'. The evaluation supports the following
symbols '*','/','+','-','exp','log','sqrt'. The evaluation uses a gadget
parameter object for its evaluation.
}
\author{
Bjarki Thor Elvarsson
}

