% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{firststep}
\alias{firststep}
\title{The first timestep}
\usage{
firststep(n, mu, sigma, l, z, numofareas, probarea, minage, maxage)
}
\arguments{
\item{n}{Number of age 1 individuals}

\item{mu}{Vector of mean length per age.}

\item{sigma}{Vector of standard deviation of length per age.}

\item{l}{Vector of length groups}

\item{z}{Natural mortality per age}

\item{numofareas}{Number of areas}

\item{probarea}{Vector of proportions living in an area.}

\item{minage}{Minimum age of the species.}

\item{maxage}{Maximum age of the species.}
}
\value{
matrix with initial distribution
}
\description{
The simulation is started with an initial population, the age of which
ranges from the minimum age of the immature stock to the maximum age
of the mature stock. This population is calculated as follows:
Let \eqn{n_{a}} be the number of age \eqn{a} individuals in the first
timestep, \eqn{\mu_{a}} the mean length at age \eqn{a} and \eqn{\sigma_{a}}
the standard deviation of length at age \eqn{a}. For the minimum age
(representing recruitment) see the recruitment function. For a given
constant mortality \eqn{Z_{0}} we get
\deqn{n_{a} = n_{a-1}e^{-Z_{0}}}
to calculate the number at age for all \eqn{a}.
The number in lengthgroup $i$ at age $a$ in timestep 1 can then be
calculated from:
\deqn{N_{i,a,1} = n_a\left(\Phi\left(\frac{l_{i+1}-\mu_a}{\sigma_a}\right)-
\Phi\left(\frac{l_{i}-\mu_a}{\sigma_a}\right)\right)}{N_{i,a,1} = n_a (Phi(l_{i+1}-mu_a)/sigma_a)-Phi(l_{i}-\mu_a)/sigma_a))}
where \eqn{l_{i}} and \eqn{l_{i+1}} are the endpoints of lengthgroup \eqn{i},
\eqn{N_{l,a,t}} is the number at age \eqn{a} in lengthgroup \eqn{l} at
timestep \eqn{t} and \eqn{\Phi}{Phi} is the probability function for Normal
distribution. 
NB: in Gadget (which is programmed in C++) the value of \eqn{\Phi}{Phi} is
approximated whereas R uses integration. To ensure
compatibility between the models, the initial population for Gadget is
entered directly from the initial population file rather than
calculated from a Normal distribution. While this is an option
within Gadget, it is not the standard method.
}
