% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{fix_headers}
\alias{fix_headers}
\title{Fix data headers in Gadget input files}
\usage{
fix_headers(main.file = "main", path = ".")
}
\arguments{
\item{main.file}{name of the main file}

\item{path}{location of the gadget model directory}
}
\description{
Recent revisions of RGadget introduced the gadgetfile class, which is a fairly flexible S3 class to interact with Gadget input files. 
To ensure that data object are read properly, i.e. as data.frame, the following header needs to be inserted to indicate that the following 
sections indicate data\cr
\cr
\code{; -- data --} \cr
\code{; col1 col2 ...}\cr
\cr
This function attempts to translate older gadget models, i.e. those not developed using Rgadget, to the new format
}
\examples{
\dontrun{
fix_headers()
}
}
