% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{fleet_suit}
\alias{fleet_suit}
\title{Write suitability line for a fleet}
\usage{
fleet_suit(fleet = "comm", stock = NULL, fun = "exponentiall50",
  params = NULL)
}
\arguments{
\item{fleet}{Character. The name of the fleet.}

\item{stock}{Character. The name of the stock that is caught.}

\item{fun}{Character. The selection function to be used by the fleet}

\item{params}{A list of named parameters that are needed by \code{fun}. 
Names will be used to output the switches to be used by gadget.
Alternatively, a vector of numbers can be specified.}
}
\value{
A character vector of the form 
  <preyname> function <functionname> <vector of parameters>
}
\description{
Write suitability line for a fleet
}
\examples{
gadgetfleet('Modelfiles/fleet', '~', missingOkay=T) \%>\%
    gadget_update('totalfleet',
                   name = 'comm',
                   suitability = fleet_suit(fleet='comm', 
                                            stock='zebrafish', 
                                            fun='newexponentiall50',
                                            params=list('alpha', 'l50')),
                   data = comm.landings[[1]])
}
\author{
Paul Frater
}
