% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gadget.bootfit}
\alias{gadget.bootfit}
\title{bootstrap fit}
\usage{
gadget.bootfit(main = "main", dparam.file = "bsres_v1.RData",
  bsprint.file = "bsprint.RData", fleet.predict = data.frame(fleet = "comm",
  ratio = 1), mat.par = NULL, .parallel = TRUE)
}
\arguments{
\item{main}{}

\item{dparam.file}{}

\item{bsprint.file}{}

\item{fleet.predict}{}

\item{mat.par}{}

\item{.parallel}{}
}
\value{
list of bootstrap fit things
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

