% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{gadget.bootforward}
\alias{gadget.bootforward}
\title{Gadget bootstrap forward}
\usage{
gadget.bootforward(years = 20, params.file = "params.final",
  main.file = "main.final", pre = "PRE", effort = 0.2,
  fleets = data.frame(fleet = "comm", ratio = 1), num.trials = 10,
  bs.wgts = "BS.WGTS", bs.samples = 1:1000, check.previous = FALSE,
  rec.window = NULL, mat.par = NULL, stochastic = TRUE,
  .parallel = TRUE)
}
\arguments{
\item{years}{}

\item{params.file}{}

\item{main.file}{}

\item{pre}{}

\item{effort}{}

\item{fleets}{}

\item{num.trials}{}

\item{bs.wgts}{}

\item{bs.samples}{}

\item{check.previous}{}

\item{rec.window}{}

\item{mat.par}{}

\item{stochastic}{}

\item{.parallel}{}
}
\value{
list of bootstrap results
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

