% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gadget.bootstrap}
\alias{gadget.bootstrap}
\title{Bootstrap control}
\usage{
gadget.bootstrap(bs.likfile = "likelihood.bs", bs.samples = 1,
  main = "main", optinfofile = "optinfo", bs.wgts = "BS.WGTS",
  bs.data = NULL, params.file = "params.in", rew.sI = FALSE,
  grouping = NULL, qsub.script = "bootstrap.sh", run.final = FALSE,
  PBS = TRUE, cv.floor = NULL)
}
\arguments{
\item{bs.likfile}{Likelihood template file}

\item{bs.samples}{number (or vector of numbers) indicating what bootstrap
samples should be used}

\item{main}{Main file for the gagdet model}

\item{optinfofile}{optinfofile used in the optimization.}

\item{bs.wgts}{folder containing the resulting reweights}

\item{bs.data}{folder containing the bootstrap dataset obtain from the DW}

\item{params.file}{}

\item{rew.sI}{should the survey indices be reweighted seperately}

\item{grouping}{list of grouped likelihood components}

\item{qsub.script}{name of the qsub script if the calculations is meant to be run on a cluser}

\item{run.final}{logical, is this the final run or weighting run.}

\item{PBS}{logical, is this a cluster run?}
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

