% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{gadget.bootypr}
\alias{gadget.bootypr}
\title{Gadget bootstrap yield per recruit}
\usage{
gadget.bootypr(params.file = "params.final", main.file = "main.final",
  effort = seq(0, 1, by = 0.01), begin = 1990, end = 2020,
  fleets = data.frame(fleet = "comm", ratio = 1), ypr = "YPR",
  bs.wgts = "BS.WGTS", bs.samples = 1:1000, .parallel = TRUE)
}
\value{
yield per recruit for the bootstrap
}
\author{
Bjarki Thor Elvarsson
}
