% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gadget.forward}
\alias{gadget.forward}
\title{Gadget forward simulation}
\usage{
gadget.forward(years = 20, params.file = "params.out", main.file = "main",
  pre = "PRE", num.trials = 10, fleets = data.frame(fleet = "comm", ratio
  = 1), biomass = FALSE, effort = 0.2, spawnmodel = "none",
  spawnvar = NULL, selectedstocks = NULL, biomasslevel = NULL,
  check.previous = FALSE, save.results = TRUE, stochastic = TRUE,
  mat.par = NULL, rec.window = NULL, compact = TRUE)
}
\arguments{
\item{years}{}

\item{params.file}{}

\item{main.file}{}

\item{pre}{}

\item{num.trials}{}

\item{fleets}{}

\item{biomass}{}

\item{effort}{}

\item{spawnmodel}{}

\item{spawnvar}{}

\item{selectedstocks}{}

\item{biomasslevel}{}

\item{check.previous}{}

\item{save.results}{}

\item{stochastic}{}

\item{mat.par}{}

\item{rec.window}{}

\item{compact}{}
}
\value{
list of results
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

