% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.optinfo}
\alias{read.gadget.optinfo}
\alias{write.gadget.optinfo}
\title{Gadget optimiser settings}
\usage{
read.gadget.optinfo(file = "optinfofile")

write.gadget.optinfo(optinfo, file = "optinfofile")
}
\arguments{
\item{file}{location of the optinfofile}

\item{optinfo}{optinfo object}

\item{file}{file}

\item{location}{location}
}
\value{
optinfo object

text of the optinfofile (if desired)
}
\description{
\code{read.gadget.optinfor} reads optinfo parameters from file

\code{write.gadget.optinfo} write optinfo to file
}
\author{
Bjarki ??r Elvarsson

Bjarki ??r Elvarsson
}
