% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetparameters.R
\name{read.gadget.parameters}
\alias{read.gadget.parameters}
\alias{write.gadget.parameters}
\title{Gadget parameters}
\usage{
read.gadget.parameters(file = "params.in")

write.gadget.parameters(params, file = "params.out")
}
\arguments{
\item{file}{a string naming the file to write to}

\item{params}{params dataframe}
}
\value{
dataframe

a string containing the text of the params file (if desired)
}
\description{
\code{read.gadget.parameters} reads a gadget parameter file
}
\author{
Bjarki  Thor Elvarsson

Bjarki Þór Elvarsson
}
