% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfit.R
\name{gadget.retro.fit}
\alias{gadget.retro.fit}
\title{Retro fit}
\usage{
gadget.retro.fit(pre = "RETRO")
}
\arguments{
\item{pre}{Folder containing the retro run}
}
\value{
gadget fit list
}
\description{
Compile the fit objects from retrospective runs
}
