% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{gadget.sensitivity}
\alias{gadget.sensitivity}
\title{Gadget sensitivity}
\usage{
gadget.sensitivity(file = "params.out", outer.range = 0.5,
  outer.stepsize = 0.05, inner.range = 0.05, inner.stepsize = 0.01,
  opt = TRUE, vars.all = TRUE, var.names = "",
  gadget.exe = "gadget", sens.in = "sens.in", lik.out = "lik.sens",
  within.bounds = TRUE, main.file = "main", sens.dir = "SENS",
  calc.full = FALSE)
}
\arguments{
\item{file}{name of the input file with the initial point}

\item{outer.range}{The outer ranges of the parameter value
considered, defined in terms of percentages.}

\item{outer.stepsize}{The increments/stepsize within in the outer range.}

\item{inner.range}{Inner range where the finer mesh should be used}

\item{inner.stepsize}{Inner stepsize.}

\item{opt}{Will we be looking at only the optimized variables, or
all of them?}

\item{vars.all}{(logical) Will we be looking at all variables or
just some?}

\item{var.names}{If only a few, which ones will they be? Can be
blank if we are using all variables}

\item{gadget.exe}{name of the gadget executable}

\item{sens.in}{name of the resulting gadget input file}

\item{lik.out}{a string containing the name of the likelihood
output file}

\item{within.bounds}{should gadget be restricted to}

\item{main.file}{string naming the gadget main file used}

\item{sens.dir}{a string naming the folder where the result and
temporary files are stored. The funciton will create the folder if
it has not already been created.}

\item{calc.full}{(USE WITH CARE) should the the full hypercube of
function values be calculated. Using this switch will increase the
computation time required exponentially.}

\item{range}{The range of the sensitivity check}

\item{stepsize}{The stepsize used}
}
\value{
results from lik.out
}
\description{
This function implements a crude sensitivity analysis of a gadget
simulation The test is run by changing each of the variables in
your model by up to +/- some percentage of the initial
value. Often a higher resolution near the optimum is desired than
is requried elsewhere.
}
\author{
Bjarki Thor Elvarsson
}
