% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetoptions.R
\name{gadget.skeleton}
\alias{gadget.skeleton}
\title{Gadget skeleton}
\usage{
gadget.skeleton(time, area, stocks, fleets = NULL)
}
\arguments{
\item{time}{list describing the time step of the model}

\item{area}{list describing the spatial grid of the model}

\item{stocks}{list describing the stocks of the model}

\item{fleets}{list describing the fleets}
}
\value{
gadget-model class
}
\description{
Build gadget model object based on simple lists
}
\details{
Gadget skeleton, as the names suggests, creates a gadget-main
object, as S4 class, that can be used to create model input files
or drive the simple Gadget-like simulator implemented in
gadget.simulate
}
\examples{
opt <- gadget.options()
gm <- gadget.skeleton(time=opt$time,area=opt$area,
                      stocks=opt$stocks,fleets=opt$fleets)
}
\author{
Bjarki Þór Elvarsson
}
