% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{gadget.variant.dir}
\alias{gadget.variant.dir}
\title{Write the changes to the model into a model variant directory}
\usage{
gadget.variant.dir(path, variant_dir = NULL, mainfile = "main")
}
\arguments{
\item{path}{Base directory to write out to}

\item{variant_dir}{A subdirectory to write any changes out to}

\item{mainfile}{The name of the variant directories' mainfile}
}
\description{
Write the changes to the model into a model variant directory
}
