% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_catchdistribution_component}
\alias{gadget_catchdistribution_component}
\title{http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:catchdist}
\usage{
gadget_catchdistribution_component(weight = 0,
  name = "catchdistribution", data_function = "sumofsquares",
  data_function_params = list(), aggregationlevel = FALSE,
  overconsumption = FALSE, epsilon = 10, data = NULL, area = NULL,
  age = NULL, length = NULL, fleetnames = c(), stocknames = c())
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:catchdist
}
