% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_catchinkilos_component}
\alias{gadget_catchinkilos_component}
\title{http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:catchinkilos}
\usage{
gadget_catchinkilos_component(
  weight = 0,
  name = "catchinkilos",
  data = NULL,
  data_function = "sumofsquares",
  epsilon = 10,
  area = NULL,
  fleetnames = c(),
  stocknames = c()
)
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:catchinkilos
}
