% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_catchstatistics_component}
\alias{gadget_catchstatistics_component}
\title{https://hafro.github.io/gadget/docs/userguide#catchstatistics}
\usage{
gadget_catchstatistics_component(weight = 0, name = "catchstatistics",
  data_function = NULL, data = NULL, area = NULL, age = NULL,
  fleetnames = c(), stocknames = c())
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
https://hafro.github.io/gadget/docs/userguide#catchstatistics
}
