% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetWriteMethods.R
\docType{methods}
\name{gadget_dir_write,list,gadget-main-method}
\alias{gadget_dir_write,list,gadget-main-method}
\title{Write gadget main}
\usage{
\S4method{gadget_dir_write}{list,`gadget-main`}(gd, x)
}
\description{
Write gadget main
}
\author{
Bjarki Thor Elvarsson
}
