% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfleet.R
\name{gadget_discard.gadgetfleet}
\alias{gadget_discard.gadgetfleet}
\title{this function removes named fleet components}
\usage{
\method{gadget_discard}{gadgetfleet}(gf, comp_name, ...)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{comp_name}{named components to remove}
}
\description{
this function removes named fleet components
}
