% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihood.R
\name{gadget_discard.gadgetlikelihood}
\alias{gadget_discard.gadgetlikelihood}
\title{this function removes named likelihood components}
\usage{
\method{gadget_discard}{gadgetlikelihood}(gf, comp_name, ...)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{comp_name}{named components to remove}
}
\description{
this function removes named likelihood components
}
