% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprint.R
\name{gadget_discard.gadgetprint}
\alias{gadget_discard.gadgetprint}
\title{this function removes named print components}
\usage{
\method{gadget_discard}{gadgetprint}(gf, comp_name)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{comp_name}{named components to remove}
}
\description{
this function removes named print components
}
