% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{gadget_discard}
\alias{gadget_discard}
\title{Discard named component}
\usage{
gadget_discard(gf, comp_name, ...)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{comp_name}{named components to remove}
}
\description{
Remove unwanted components from file
}
