% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_forward.R
\name{gadget_evaluate}
\alias{gadget_evaluate}
\title{Evaluate a gadget model}
\usage{
gadget_evaluate(
  path = ".",
  params.in = NULL,
  params.out = NULL,
  lik.out = NULL,
  ...
)
}
\arguments{
\item{path}{locatoin of the gadget model}

\item{params.in}{parameter input file}

\item{params.out}{parameter output file}

\item{lik.out}{likelihood output file}

\item{...}{additional arguments passed on to callGadget}
}
\value{
path
}
\description{
Evaluate a gadget model
}
