% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\docType{data}
\name{gadget_filetypes}
\alias{gadget_filetypes}
\title{Recognised GADGET file types and their quirks}
\description{
\describe{
    \item{file_type}{The name of the file type}
    \item{mainfile_section}{Writing a file of this type will cause the corresponding mainfile entry to be updated}
    \item{bare_component}{Keys matching this regex (or every component, if TRUE) do not have brackets around their names}
    \item{implicit_component}{Keys matching this regex are treated as if the are implicitly broken up into components}
}
}
