% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_fit.R
\name{gadget_fit}
\alias{gadget_fit}
\title{Gadget fit}
\usage{
gadget_fit(
  gd,
  params.in = attr(gd, "params_in"),
  fit = "FIT",
  f.age.range = NULL,
  steps = 1
)
}
\arguments{
\item{gd}{gadget.variant.dir}

\item{params.in}{input parameters}

\item{fit}{location of the fit folder}

\item{f.age.range}{age range for F}

\item{steps}{(unused)}
}
\value{
list
}
\description{
Gadget fit
}
