% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_likelihoodprinter_component}
\alias{gadget_likelihoodprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:likelihoodprinter}
\usage{
gadget_likelihoodprinter_component(printfile, likelihoodcomponent)
}
\arguments{
\item{printfile}{Name for the output file to be created}

\item{likelihoodcomponent}{name of the likelihood component}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:likelihoodprinter
}
