% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_likelihoodsummaryprinter_component}
\alias{gadget_likelihoodsummaryprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:likelihoodsummaryprinter}
\usage{
gadget_likelihoodsummaryprinter_component(printfile)
}
\arguments{
\item{printfile}{Name for the output file to be created}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:likelihoodsummaryprinter
}
