% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGadget.R
\name{gadget_optimize}
\alias{gadget_optimize}
\title{Evaluate a gadget model}
\usage{
gadget_optimize(
  path = ".",
  params.in = NULL,
  params.out = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{path}{locatoin of the gadget model}

\item{params.in}{parameter input file}

\item{params.out}{parameter output file}

\item{control}{optimization parameters, can be the location of the optinfo file, gadgetfile object or NULL}

\item{...}{additional arguments passed on to callGadget}
}
\value{
path
}
\description{
Evaluate a gadget model
}
\examples{
\dontrun{
gd <- gadget.variant.dir('model_dir')
optinfo <- read.gadget.file('/',system.file('extdata','optinfo',package = 'Rgadget'))
gadget_optimize(gd,params.in = 'params.in',params.out = 'params.opt', control = optinfo)
}
}
