% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_recaptures_component}
\alias{gadget_recaptures_component}
\title{http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:recaptures}
\usage{
gadget_recaptures_component(weight = 0, name = "recaptures", data = NULL)
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:recaptures
}
