% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_recstatistics_component}
\alias{gadget_recstatistics_component}
\title{http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:recstat}
\usage{
gadget_recstatistics_component(weight = 0, name = "recstatistics", data = NULL)
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}

\item{data}{data.frame}
}
\description{
http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:recstat
}
