% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_stockdistribution_component}
\alias{gadget_stockdistribution_component}
\title{https://hafro.github.io/gadget/docs/userguide#stockdistribution}
\usage{
gadget_stockdistribution_component(weight = 0,
  name = "stockdistribution", data_function = "sumofsquares",
  overconsumption = FALSE, epsilon = 10, data = NULL, area = NULL,
  age = NULL, length = NULL, fleetnames = c(), stocknames = c())
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
https://hafro.github.io/gadget/docs/userguide#stockdistribution
}
