% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_stockfullprinter_component}
\alias{gadget_stockfullprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockfullprinter}
\usage{
gadget_stockfullprinter_component(
  printfile,
  stockname,
  precision = 1,
  printatstart = 1,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{A descriptive name for the component, default same as component type}

\item{stockname}{A numeric weighting}

\item{precision}{}

\item{printatstart}{}

\item{yearsandsteps}{}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockfullprinter
}
