% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_stockpreyfullprinter_component}
\alias{gadget_stockpreyfullprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockpreyfullprinter}
\usage{
gadget_stockpreyfullprinter_component(
  printfile,
  preynames,
  area,
  len,
  precision = 1,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{A descriptive name for the component, default same as component type}

\item{precision}{}

\item{yearsandsteps}{}

\item{stockname}{A numeric weighting}

\item{printatstart}{}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockpreyfullprinter
}
