% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_stockpreyprinter_component}
\alias{gadget_stockpreyprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockpreyprinter}
\usage{
gadget_stockpreyprinter_component(
  printfile,
  preynames,
  area,
  age,
  len,
  precision = 15,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{Name for the output file to be created}

\item{preynames}{Names of the prey stocks}

\item{area}{list of area aggregations}

\item{age}{list of age aggregations}

\item{len}{list of length aggregations}

\item{precision}{optional value defining numerical precision}

\item{yearsandsteps}{character defining when to print, defaults to "all all"}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockpreyprinter
}
