% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_stockprinter_component}
\alias{gadget_stockprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockprinter}
\usage{
gadget_stockprinter_component(
  printfile,
  stocknames,
  area,
  age,
  len,
  precision = 1,
  printatstart = 1,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{A descriptive name for the component, default same as component type}

\item{precision}{}

\item{printatstart}{}

\item{yearsandsteps}{}

\item{stockname}{A numeric weighting}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockprinter
}
