% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_stockstdprinter_component}
\alias{gadget_stockstdprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockstdprinter}
\usage{
gadget_stockstdprinter_component(
  printfile,
  stockname,
  scale = 1,
  precision = 1,
  printatstart = 1,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{A descriptive name for the component, default same as component type}

\item{stockname}{A numeric weighting}

\item{scale}{The datafile lists these weights and the power that is to be used for each parameter}

\item{precision}{}

\item{printatstart}{}

\item{yearsandsteps}{}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockstdprinter
}
