% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_surveydistribution_component}
\alias{gadget_surveydistribution_component}
\title{https://hafro.github.io/gadget/docs/userguide#surveydistribution}
\usage{
gadget_surveydistribution_component(weight = 0,
  name = "surveydistribution", data = NULL, area = NULL,
  length = NULL, age = NULL, stocknames = c(),
  fittype = "linearfit", parameters = NULL, suitability = list(),
  slope = NULL, intercept = NULL, epsilon = 10,
  likelihoodtype = "multinomial")
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
https://hafro.github.io/gadget/docs/userguide#surveydistribution
}
