% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgettime.R
\name{gadget_update.gadgettime}
\alias{gadget_update.gadgettime}
\title{Update gadget time components in a time file}
\usage{
\method{gadget_update}{gadgettime}(gf, ...)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{...}{Valid options for a time file; firstyear, firststep, lastyear, laststep, notimesteps. See gadget user guide}
}
\description{
Replace and/or append new time comonents to an existing file
}
\examples{
library(magrittr)  # import \%>\% function
path <- './model'
gadgettime('time', path, missingOkay = TRUE) \%>\%
   gadget_update(firstyear = 1990, lastyear = 2000) \%>\% 
   write.gadget.file(path)
}
