% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{gadget_update}
\alias{gadget_update}
\title{Update gadget components}
\usage{
gadget_update(gf, component, ...)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{component}{The component to update}

\item{...}{Keys to update.}
}
\description{
Update a component within this file.
}
\details{
For more information, see a specific implementation:
\enumerate{
  \item \link{gadget_update.gadgetstock}
  \item \link{gadget_update.gadgetlikelihood}
}
}
