% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{gadgetdata}
\alias{gadgetdata}
\title{Construct a data-only gadget file}
\usage{
gadgetdata(file_name, data)
}
\arguments{
\item{file_name}{Filename the output should be written to / read from}

\item{data}{A data.frame that the file should contain}
}
\value{
A gadgetfile S3 class, use \code{write.gadget.file} to write it to disk
}
\description{
Construct a data-only gadget file
}
