% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{gadgetfile}
\alias{gadgetfile}
\title{Construct a new gadgetfile S3 object}
\usage{
gadgetfile(file_name, file_type = "generic", components = list())
}
\arguments{
\item{file_name}{Filename the output should be written to / read from}

\item{file_type}{What sort of gadget file this is, e.g. 'time', 'area', ...
See \code{Rgadget::gadget_filetypes} for recognised types.
Setting this will enable any quirks of that particular file type,}

\item{components}{A list of lists, representing each component.}
}
\value{
A gadgetfile S3 class, use \code{write.gadget.file} to write it to disk
}
\description{
Constructor for objects representing a gadget model input
file.
}
\details{
gadgetfile objects break all files down into the following form:
\preformatted{; preamble before the first unnamed component
key	value
; preamble before the second named component
[component2_name]
key1	value
key2	value
; preamble before the second named component
[component3_name]
key	value
  . . .
; preamble before the data component
[data_component]
; -- data --
; a	b	c
1	2	3
}

Note that values can also be \code{gadgetfile} objects, in this case
That gadgetfile is written out separately, and value is replaced with
the filename of the sub-file.
}
\examples{
gadgetfile("cod.likelihood", "likelihood",
   components = list(
       component = list(type = "penalty", aggfile = gadgetfile("aggfile", "areaagg", components = list()))
   ))
}
