% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.likelihood}
\alias{read.gadget.likelihood}
\alias{write.gadget.likelihood}
\alias{get.gadget.likelihood}
\alias{read.gadget.main}
\alias{write.gadget.main}
\alias{clear.spaces}
\alias{read.gadget.data}
\alias{strip.comments}
\alias{read.gadget.wgts}
\alias{read.gadget.grouping}
\title{Old style gadget file input and output (mostly deprecated)}
\usage{
read.gadget.likelihood(files = "likelihood")

write.gadget.likelihood(
  lik,
  file = "likelihood",
  data.folder = NULL,
  bs.sample = NULL
)

get.gadget.likelihood(likelihood, comp, inverse = FALSE)

read.gadget.main(file = "main")

write.gadget.main(main, file = "main")

clear.spaces(text)

read.gadget.data(likelihood, debug = FALSE, year_range = NULL)

strip.comments(file = "main")

read.gadget.wgts(
  params.file = "params.in",
  wgts = "WGTS",
  likelihood = "likelihood",
  lik.pre = "lik.",
  params.pre = "params.",
  parallel = FALSE
)

read.gadget.grouping(lik = read.gadget.likelihood(), wgts = "WGTS")
}
\arguments{
\item{files}{a vector of character strings containing the names of the likelihood files}

\item{lik}{old style gadget likelihood object}

\item{file}{location of the gadget input file}

\item{data.folder}{location of data folder (if changed)}

\item{bs.sample}{(for bootstrap), appends the appropriate replicate number to data file}

\item{likelihood}{likelihood file}

\item{comp}{selected likelihood components}

\item{inverse}{(logical) should inverse selection be applied}

\item{main}{gadget.main object}

\item{text}{string}

\item{debug}{should debug information be printed}

\item{year_range}{limit the years read in}

\item{params.file}{base parameter file}

\item{wgts}{logcation of the WGTS folder}

\item{lik.pre}{strings matching the likelihood output}

\item{params.pre}{strings matching the parameter estimates}

\item{parallel}{should the files be read in parallel}
}
\value{
object of class gadget.likelihood, i.e. a list containing the various likelihood components

character string corresponding to the likelihood file (if desired)

likelihood object

object of class gadget.main

text of the main file (if desired)

list or matrix containing the (non-empty) values from the string

list of dataframes and degress of freedom

list containing the lines from the file stripped of unwanted text.

data.frame with parameter estimates and likelihood output from the iterative reweighting folder.

list of wgts groupings
}
\description{
\code{read.gadget.likelihood} reads the likelihood (input) file for gadget. The format of
the likelihood file is described in gadget's user manual.

\code{read.gadget.data} reads data used by the various components

\code{read.gadget.wgts} reads the output from iterative weighting likelihood output

\code{read.gadget.grouping} reads the the likelihood grouping from the WGTS folder
}
\author{
Bjarki Þór Elvarsson

Bjarki Þór Elvarsson

Bjarki Thor Elvarsson

Bjarki Þór Elvarsson

Bjarki Þór Elvarsson

Bjarki Þór Elvarsson

Bjarki Þór Elvarsson

Bjarki Thor Elvarsson
}
