% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.wgts}
\alias{read.gadget.wgts}
\alias{merge.formula}
\alias{eval.gadget.formula}
\alias{read.gadget.grouping}
\title{gadget.fit helper functions}
\usage{
read.gadget.wgts(
  params.file = "params.in",
  wgts = "WGTS",
  likelihood = "likelihood",
  lik.pre = "lik.",
  params.pre = "params.",
  parallel = FALSE
)

\method{merge}{formula}(txt)

eval.gadget.formula(gad.for, par = data.frame())

read.gadget.grouping(lik = read.gadget.likelihood(), wgts = "WGTS")
}
\arguments{
\item{params.file}{base parameter file}

\item{wgts}{logcation of the WGTS folder}

\item{likelihood}{likelihood file}

\item{lik.pre}{strings matching the likelihood output}

\item{params.pre}{strings matching the parameter estimates}

\item{parallel}{should the files be read in parallel}

\item{txt}{gadget formula text}

\item{gad.for}{gadget formula}

\item{par}{gadget parameters object}

\item{lik}{old style gadget likelihood object}
}
\value{
data.frame with parameter estimates and likelihood output from the iterative reweighting folder.

txt where the formula has been properly compiled

a vector of evaluated gadget formulas

list of wgts groupings
}
\description{
\code{read.gadget.wgts} reads the output from iterative weighting likelihood output

\code{merge.formula} merges txt to gadget.formula

\code{eval.gadget.formula} Evaluate gadget formulas, which are in reverse polish notation, ie
'(* x y)' which is equivalent to 'x*y'. The evaluation supports the following
symbols '*','/','+','-','exp','log','sqrt'. The evaluation uses a gadget
parameter object for its evaluation.

\code{read.gadget.grouping} reads the the likelihood grouping from the WGTS folder
}
\author{
Bjarki Thor Elvarsson
}
