% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfleet.R
\name{gadgetfleet}
\alias{gadgetfleet}
\title{Create a gadgetfleet object}
\usage{
gadgetfleet(file_name, path, missingOkay = FALSE)
}
\arguments{
\item{file_name}{The name of the fleet file}

\item{path}{The path to the gadget directory to read from}

\item{missingOkay}{If \code{TRUE}, return an empty fleet file object if file does not exist.}
}
\value{
A list of fleet components representing file
}
\description{
Create a fleet file object, from fresh or an existing file.
}
\examples{
path <- './cod-model'
gadgetfleet('fleet', path, missingOkay = TRUE)  # Read 'fleet' fleet file, creating it if it doesn't exist
}
