% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihood.R
\name{gadgetlikelihood}
\alias{gadgetlikelihood}
\title{Create a gadgetlikelihood object}
\usage{
gadgetlikelihood(file_name, path, missingOkay = FALSE)
}
\arguments{
\item{file_name}{The name of the likelihood file}

\item{path}{The path to the gadget directory to read from}

\item{missingOkay}{If \code{TRUE}, return an empty likelihood file object if file does not exist.}
}
\value{
A list of likelihood components representing file
}
\description{
Create a likelihood file object, from fresh or an existing file.
}
\examples{
path <- './cod-model'
gadgetlikelihood('likelihood', path, missingOkay = TRUE)  # Read 'likelihood' likelihood file, creating it if it doesn't exist
}
