% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadgetprintcomponent}
\alias{gadgetprintcomponent}
\title{Wrapper to choose a component by name}
\usage{
gadgetprintcomponent(component, ...)
}
\arguments{
\item{component}{Call the corresponding gadget_(component)_component function}

\item{...}{Arguments to pass to it}
}
\description{
Wrapper to choose a component by name
}
