% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprint.R
\name{gadgetprintfile}
\alias{gadgetprintfile}
\title{Create a gadgetprint object}
\usage{
gadgetprintfile(file_name, path, missingOkay = FALSE)
}
\arguments{
\item{file_name}{The name of the print file}

\item{path}{The path to the gadget directory to read from}

\item{missingOkay}{If \code{TRUE}, return an empty print file object if file does not exist.}
}
\value{
A list of print components representing file
}
\description{
Create a print file object, from fresh or an existing file.
}
\examples{
\dontrun{
path <- './cod-model'
# Read 'print' print file, creating it if it doesn't exist
gadgetprintfile('print', path, missingOkay = TRUE)  
}
}
