% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetstock.R
\name{gadgetstock}
\alias{gadgetstock}
\title{Create a gadgetstock object}
\usage{
gadgetstock(stock_name, path, missingOkay = FALSE)
}
\arguments{
\item{stock_name}{The name of the stock, as}

\item{path}{The path to the gadget directory to read from}

\item{missingOkay}{If \code{TRUE}, return an empty gadgetstock object if file does not exist.}
}
\value{
A list of lists representing the gadget stock file
}
\description{
Create a gadgetstock object, from fresh or an existing file
}
\examples{
path <- './cod-model'
gadgetstock('codmat', path, missingOkay = TRUE)  # Read codmat from './cod-model', throw an error if it doesn't exist
gadgetstock('codimm', path, missingOkay = TRUE)  # Read codimm stock / create a new one
}
