% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgettime.R
\name{gadgettime}
\alias{gadgettime}
\title{Create a gadgettime object}
\usage{
gadgettime(file_name, path, missingOkay = FALSE)
}
\arguments{
\item{file_name}{The name of the time file}

\item{path}{The path to the gadget directory to read from}

\item{missingOkay}{If \code{TRUE}, return an empty time file object if file does not exist.}
}
\value{
A list of time components representing file
}
\description{
Create a time file object, from fresh or an existing file.
}
\examples{
path <- './cod-model'
gadgettime('Modelfiles/time', path, missingOkay = TRUE)  # Read 'Modelfiles/time' time file, creating it if it doesn't exist
}
