% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{get.gadget.growth}
\alias{get.gadget.growth}
\title{get gadget growth}
\usage{
get.gadget.growth(stocks, params, dt = 0.25, age.based = FALSE,
  recl = FALSE)
}
\arguments{
\item{stocks}{gadget.stock}

\item{params}{gadget.parms}

\item{dt}{delta t}

\item{age.based}{logical}

\item{recl}{recruitment length}
}
\value{
growth matrix
}
\author{
Bjarki Thor Elvarsson
}
