% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.gadget.likelihood}
\alias{get.gadget.likelihood}
\title{Get Gagdget likelihood component(s)}
\usage{
get.gadget.likelihood(likelihood, comp, inverse = FALSE)
}
\arguments{
\item{likelihood}{likelihood object}

\item{comp}{selected likelihood components}

\item{inverse}{(logical) should inverse selection be applied}
}
\value{
likelihood object
}
\description{
This function retrives selected parts of the likelihood object
}
\author{
Bjarki Thor Elvarsson
}

