% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{get.gadget.suitability}
\alias{get.gadget.suitability}
\title{get gadget suitability}
\usage{
get.gadget.suitability(fleets, params, lengths, normalize = FALSE)
}
\value{
suitability function for a fleet
}
\author{
Bjarki Thor Elvarsson
}
