% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{growthprob}
\alias{growthprob}
\title{Update length}
\usage{
growthprob(lt, beta, lsup, k, dt, lengthgrouplen, binn)
}
\arguments{
\item{lt}{Vector of midpoints for the length groups.}

\item{beta}{Beta for the Von Bertanlanffy curve}

\item{lsup}{\eqn{l_\infty}{l_infty} for the Von Bertanlanffy.}

\item{k}{\eqn{\kappa}{kappa} for the Von Bertanlanffy.}

\item{dt}{Length of the time interval.}

\item{lengthgrouplen}{length of the lengthgroups.}

\item{binn}{is the maximum updating length.}
}
\value{
a matrix where the index (j,i) repsents the probability of going
lengthgroup i to lengthgroup j
}
\description{
Growth is according to a von Bertalanffy equation 
\deqn{\mu_{a} = L_{\infty}(1-e^{-\kappa a})}
with the lengthvbsimple growth function from Gadget implemented.
For a fish of age a and length l, mean length growth \eqn{\Delta L} is
then calculated as:
\deqn{\Delta L =L_{\infty}(1 - \frac{l}{L_{\infty}})(1 - e^{-\kappa \Delta t})}
\deqn{Delta L =L_{infty}(1 - l/L_{infty})(1 - e^{-kappa Delta t})}
The length distribution is updated using the beta-binomial
distribution, ie the probability of growing x lengthgroups, given
maximum lengthgroupgrowth n, is
\deqn{P[X = x] =  \frac{\Gamma(n+1)}{\Gamma(n-x+1)\Gamma(x+1)} \frac{\Gamma(\alpha + \beta)}
{\Gamma(n+\alpha+\beta)} \frac{\Gamma(n-x+\beta)}{\Gamma(\beta)} \frac{\Gamma(x+a)}{\Gamma(\alpha)}}
with \eqn{\alpha = \frac{\beta\Delta L}{n-\Delta L}} to preserve the mean
lengthgrowth according to the equation equation above. NB: the
expected value of \eqn{\Delta L} should be taken into consideration when
fixing n. 
Let \eqn{G = [g_{ij}]} be the length updating matrix where \eqn{g_{ij}} is the
probability of growing from lengthgroup i to lengthgroup j
obtained from the equation above.
\deqn{N_{l,a+1,t+\Delta t} = \sum_{l'\leq l}g_{l'l}N_{l,a,t}}
with \eqn{N_{l,a,t}} as described for the initial population for a >
min a .
}
