% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{init_guess}
\alias{init_guess}
\title{Initial parameter guess}
\usage{
init_guess(dat, pattern, value = 0, lower = -999, upper = 999,
  optimise = 0)
}
\arguments{
\item{dat}{Input data frame, read in from read.gadget.parameters}

\item{pattern}{name of parameter, can include grep-able wildcards}

\item{value}{new parameter value}

\item{lower}{new lower bound}

\item{upper}{new upper bound}

\item{optimise}{should the parameter be optimised}
}
\value{
updated data frame with the new parameter values
}
\description{
This is a convenience function that allows the editing of parameter files by variable (switch) names.
}
