% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{ldist}
\alias{ldist}
\title{Length distributions}
\arguments{
\item{stock.dat}{}

\item{sigma}{}

\item{dl}{}
}
\value{
data.frame
}
\description{
Length distributions
}
\author{
Bjarki Thor Elvarsson
}
