% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{length.at.age}
\alias{length.at.age}
\title{Length at age}
\usage{
\method{length}{at.age}(sim)
}
\arguments{
\item{sim}{The results from a Rgadget simulation}
}
\value{
a dataframe containin the length at age for both fleets
}
\description{
Simulated length at age
}
\author{
Bjarki Þór Elvarsson
}
