% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lengthDist}
\alias{lengthDist}
\title{Length Dist}
\usage{
lengthDist(sim, sigma = 0)
}
\arguments{
\item{sim}{Results from a Rgadget simulation}

\item{sigma}{sigma for a lognormal noise}
}
\value{
Dataframe containing the fleet length distribution.
}
\description{
Calculate the length distribution from the fleet by length groups and time.
}
\author{
Bjarki Þór Elvarsson
}

