% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{make.gadget.printfile}
\alias{make.gadget.printfile}
\title{Make gadget printfile}
\usage{
make.gadget.printfile(
  main.file = "main",
  file = "printfile",
  printatstart = 1,
  steps = 1,
  recruitment_step_age = NULL,
  gd = list(dir = ".", output = "out", aggfiles = "print.aggfiles")
)
}
\arguments{
\item{main.file}{location of the main file}

\item{file}{name of resulting printfile}

\item{printatstart}{print at the end or start of a step}

\item{steps}{print steps}

\item{recruitment_step_age}{data frame with stock and recruitment age}

\item{gd}{location of the gadget model}
}
\value{
gadget.mainfile object
}
\description{
Make gadget printfile
}
\author{
Bjarki Thor Elvarsson
}
