% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{merge.gadget.likelihood}
\alias{merge.gadget.likelihood}
\title{Merge gadget likelihood}
\usage{
\method{merge}{gadget.likelihood}(lik1, lik2)
}
\arguments{
\item{lik1}{likelihood object}

\item{lik2}{likelihood object}
}
\value{
merged gadget likelihood object
}
\description{
This function merges to likelihood objects in one
}
\author{
Bjarki Thor Elvarsson
}

