% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetformulae.R
\name{parse.gadget.formulae}
\alias{parse.gadget.formulae}
\title{Parse a GADGET formulae string}
\usage{
parse.gadget.formulae(input_formulae)
}
\arguments{
\item{input_formulae}{A single formulae string, e.g. "(+ 2 (log (- #moo 1)))"}
}
\value{
An R expression representing the gadget formulae, which could be run with \code{eval(.., list(moo = 3))}
}
\description{
Parse a GADGET formulae string
}
