% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetPlots.R
\name{plot.gadget.fit}
\alias{plot.gadget.fit}
\title{plot gadget fit}
\usage{
\method{plot}{gadget.fit}(fit, data = "sidat", type = "direct",
  dat.name = NULL)
}
\arguments{
\item{fit}{results from gadget fit}

\item{data}{what results should be plotted}
}
\value{
ggplot object
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Þór Elvarsson
}

