% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{plot.gadget.sens}
\alias{plot.gadget.sens}
\title{Plot likelihood sensitivity}
\arguments{
\item{sens}{resulsts from gadget sensitivity}

\item{comp}{what component to plot, defaults to overall score}
}
\value{
ggplot object
}
\description{
Plot the likelihood profile plot by parameter for a particular likelihood component
}
