% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{plot.gadget.ypr}
\alias{plot.gadget.ypr}
\title{Plot yield per recruit}
\usage{
\method{plot}{gadget.ypr}(ypr)
}
\arguments{
\item{ypr}{gadget.ypr object}
}
\value{
ggplot object
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

