% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{plot.gadgetSens}
\alias{plot.gadgetSens}
\title{Plot likelihood sensitivity}
\usage{
\method{plot}{gadgetSens}(sens, comp = "score")
}
\arguments{
\item{sens}{resulsts from gadget sensitivity}

\item{comp}{what component to plot, defaults to overall score}
}
\value{
ggplot object
}
\description{
Plot the likelihood profile plot by parameter for a particular likelihood component
}
