% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{pred_suit}
\alias{pred_suit}
\title{Write suitability line for a predator}
\usage{
pred_suit(pred = NA, prey = NA, fun = "newexponentiall50",
  params = NULL)
}
\arguments{
\item{pred}{Character. The name of the predator}

\item{prey}{Character. The preyname of the stock to be eaten}

\item{fun}{Character. The selection function for the predator}

\item{params}{A list of named parameters that are needed by \code{fun}. 
Names will be used to output the switches to be used by gadget.
Alternatively, a vector of numbers can be used.}
}
\value{
A character vector of the form 
  <preyname> function <functionname> <vector of parameters>
}
\description{
Write suitability line for a predator
}
\examples{
gadgetstock('garfish', '~', missingOkay=T) \%>\%
    gadget_update('doeseat',
                   name = 'comm',
                   suitability = pred_suit(pred='garfish',
                                            prey='zebrafish', 
                                            fun='newexponentiall50',
                                            params=list('alpha', 'l50')),
                   data = garfish.consumption[[1]])
}
\author{
Paul Frater
}
