% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{print.gadgetfile}
\alias{print.gadgetfile}
\title{Print given gadgetfile to stdout}
\usage{
\method{print}{gadgetfile}(x, ...)
}
\arguments{
\item{x}{gadgetfile object to print}

\item{...}{unused}
}
\description{
Print given gadgetfile to stdout
}
