% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.bootprint}
\alias{read.gadget.bootprint}
\title{read bootprint}
\usage{
read.gadget.bootprint(bs.wgts = "BS.WGTS", bs.samples = 1:1000,
  printfile = "printfile", final = "final")
}
\value{
print output for all bootstrap samples
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}
