% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.bootstrap}
\alias{read.gadget.bootstrap}
\title{read gadget bootstrap}
\usage{
read.gadget.bootstrap(params.file = "params.in", bs.wgts = "BS.WGTS",
  bs.samples = 1:100, bs.lik = "likelihood", lik.pre = "lik.",
  params.pre = "params.", parallel = FALSE)
}
\value{
wgts from all bootstrap samples
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}
