% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{read.gadget.file}
\alias{read.gadget.file}
\title{Read gadget input file, return gadgetfile S3 class representing file}
\usage{
read.gadget.file(path, file_name, file_type = "generic",
  fileEncoding = "UTF-8", missingOkay = FALSE, recursive = TRUE)
}
\arguments{
\item{path}{Base directory to read file from}

\item{file_name}{File name, or path relative to (path), to read}

\item{file_type}{What sort of gadget file this is, e.g. 'time', 'area', ...
See \code{Rgadget::gadget_filetypes} for recognised types}

\item{fileEncoding}{Character encoding of file, defaults to "UTF-8"}

\item{missingOkay}{If \code{TRUE}, return an empty gadgetfile object if file does not exist.}

\item{recursive}{Read in all nested files too (default TRUE)?}
}
\description{
Read gadget input file, return gadgetfile S3 class representing file
}
