% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.fleet}
\alias{read.gadget.fleet}
\title{read gadget fleet}
\usage{
read.gadget.fleet(fleet.file = "fleet")
}
\arguments{
\item{fleet.file}{location of the fleet file}
}
\value{
fleet file object
}
\author{
Bjarki Þór Elvarsson
}
