% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.lik.out}
\alias{read.gadget.lik.out}
\title{Read gadget lik.out}
\usage{
read.gadget.lik.out(file = "lik.out", suppress = FALSE)
}
\arguments{
\item{file}{string containing the name of the file}
}
\value{
a list containing the swicthes (names of variable), weigths
(l?kelihood components) and data (dataframe with the parameter values,
likelihood component values and the final score.
}
\description{
Read in the gadget likelihood output.
}
\author{
Bjarki Thor Elvarsson, Hoskuldur Bjornsson
}
