% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.likelihood}
\alias{read.gadget.likelihood}
\title{Read likelihood}
\usage{
read.gadget.likelihood(files = "likelihood")
}
\arguments{
\item{files}{a vector of character strings containing the names of the likelihood files}
}
\value{
object of class gadget.likelihood, i.e. a list containing the various likelihood components
}
\description{
This functions reads the likelihood (input) file for gadget. The format of
the likelihood file is described in gadget's user manual.
}
\author{
Bjarki ??r Elvarsson
}
