% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.main}
\alias{read.gadget.main}
\title{Read main}
\usage{
read.gadget.main(file = "main")
}
\arguments{
\item{file}{main file location}
}
\value{
object of class gadget.main
}
\description{
Read gadget's main file
}
\author{
Bjarki ??r Elvarsson
}

