% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gadget.model}
\alias{read.gadget.model}
\title{read gadget main}
\usage{
read.gadget.model(main.file = "main", model.name = "Gadget-model")
}
\arguments{
\item{main.file}{location of the main file}
}
\value{
mainfile object
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

