% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gadget.optinfo}
\alias{read.gadget.optinfo}
\title{Read gadget}
\usage{
read.gadget.optinfo(file = "optinfofile")
}
\arguments{
\item{file}{location of the optinfofile}
}
\value{
optinfo object
}
\description{
Read optinfo parameters from file
}
\author{
Bjarki ??r Elvarsson
}

