% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gadget.parameters}
\alias{read.gadget.parameters}
\title{Read param}
\usage{
read.gadget.parameters(file = "params.in")
}
\arguments{
\item{file}{parameter file}
}
\value{
dataframe
}
\description{
Read gadget parameter file
}
\author{
Bjarki  Thor Elvarsson
}

