% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.printfile}
\alias{read.gadget.printfile}
\title{Read gadget printfile}
\usage{
read.gadget.printfile(file = "printfile")
}
\arguments{
\item{file}{string containing the path to the printfile}
}
\value{
list of the prinfile components.
}
\description{
Read gadget printfile
}
\author{
Bjarki Thor Elvarsson
}
