% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.results}
\alias{read.gadget.results}
\title{Gadget results}
\usage{
read.gadget.results(grouping = list(), final = list(final = "final"),
  wgts = "WGTS", normalize = FALSE)
}
\arguments{
\item{grouping}{--defunct--}

\item{final}{--defunct--}

\item{wgts}{location of the folder whith results from the
iterative refweighting}

\item{normalize}{(logical) should the resulting table be normalized}
}
\value{
table with SS
}
\description{
Produce diagnostics
}
\author{
Bjarki Thor Elvarsson
}
