% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.stockfiles}
\alias{read.gadget.stockfiles}
\title{read gadget stock}
\usage{
read.gadget.stockfiles(stock.files)
}
\arguments{
\item{stock.files}{}
}
\value{
list of gadget-stock objects
}
\description{
Read in Gadget stockfiles
}
\author{
Bjarki Thor Elvarsson
}

