% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.wgts}
\alias{read.gadget.wgts}
\title{read wgts}
\usage{
read.gadget.wgts(params.file = "params.in", wgts = "WGTS",
  likelihood = "likelihood", lik.pre = "lik.", params.pre = "params.",
  parallel = FALSE)
}
\arguments{
\item{params.file}{base parameter file}

\item{wgts}{location of the reweighting folder}

\item{likelihood}{likelihood file}

\item{lik.pre}{strings matching the likelihood output}

\item{params.pre}{strings matching the parameter estimates}

\item{parallel}{should the files be read in parallel}
}
\value{
data.frame with parameter estimates and likelihood output from the iterative reweighting folder.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

