% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.wgtsprint}
\alias{read.gadget.wgtsprint}
\title{read wgts print}
\usage{
read.gadget.wgtsprint(wgts = "WGTS", comp = "final", out.pre = "out.",
  parallel = FALSE)
}
\arguments{
\item{wgts}{location of the WGTS folder}

\item{comp}{likelihood component}

\item{out.pre}{output}

\item{parallel}{logical}
}
\value{
printfiles from the wgts
}
\author{
Bjarki Thor Elvarsson
}
