% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.printfiles}
\alias{read.printfiles}
\title{Read gadget printfiles}
\usage{
read.printfiles(path = ".", suppress = FALSE)
}
\arguments{
\item{path}{a character string with the name of the folder containing the printfiles}

\item{suppress}{logical should warnings of missing files be suppressed}
}
\value{
a list containing the data that has been read in named after the files found in path.
}
\description{
This function reads in the gadget output files defined in
printfiles. This is a quick and dirty implementation that has been
designed to read a few nice examples, so it may work for some instances
and it may not. It assumes that the last line containing a comment ';'
is the line describing the column names and is the last comment line.
}
