% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strip.comments}
\alias{strip.comments}
\title{Strip comments}
\usage{
strip.comments(file = "main")
}
\arguments{
\item{file}{location of the gadget input file}
}
\value{
list containing the lines from the file stripped of unwanted text.
}
\description{
Helper function created to clear out all comments (indicated by ';') and
unwanted spaces from gadget input and output files.
}
\author{
Bjarki Thor Elvarsson
}

