% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.gadget.options}
\alias{summary.gadget.options}
\title{Summary of gadget.options}
\usage{
\method{summary}{gadget.options}(opt)
}
\arguments{
\item{opt}{gadget.options list}
}
\description{
summary of the simulation defined by gadget.options
}

