% Generated by roxygen2 (4.0.2): do not edit by hand
\name{survey.index}
\alias{survey.index}
\title{Survey indices}
\usage{
survey.index(stock.dat, split, sigma = 0, alpha = 1, beta = 1)
}
\arguments{
\item{sigma}{sigma for a log-normal noise for the indicies}

\item{sim}{Results from a Rgadget simulation}
}
\value{
Dataframe with the survey indices
}
\description{
Calculates the survey indices for the simulated stocks.
}
\author{
Bjarki Þór Elvarsson
}

