% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{surveydist_suit}
\alias{surveydist_suit}
\title{Write suitability line for the surveydistribution likelihood component}
\usage{
surveydist_suit(survey.name = "survey", stock = NULL,
  fun = "newexponentiall50", params = NULL)
}
\arguments{
\item{survey.name}{Character. The name of the survey}

\item{stock}{Character. The name of the stock surveyed}

\item{fun}{Character. The selection function for the survey}

\item{params}{A list of named parameters that are needed by \code{fun}. 
Names will be used to output the switches to be used by gadget.
Alternatively, a vector of numbers can be used.}
}
\value{
A character vector of the form 
  function <functionname> <vector of parameters>
}
\description{
Write suitability line for the surveydistribution likelihood component
}
\examples{
gadgetlikelihood('likelihood', '~', missingOkay=T) \%>\%
   gadget_update('surveydistribution',
                 name = 'ldist.spr',
                 weight = 1,
                 data = ldist.igfs[[1]],
                 parameters = quote(exp(spr.si.beta)) \%>\%
                               to.gadget.formulae(),
                 suitability = surveydist_suit(stock = 'zebrafish',
                                             fun = 'constant',
                                             params = 1),
                 stocknames = 'zebrafish')
}
\author{
Paul Frater
}
