% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetformulae.R
\name{to.gadget.formulae}
\alias{to.gadget.formulae}
\title{Turn R expression into GADGET formulae string}
\usage{
to.gadget.formulae(ex)
}
\arguments{
\item{ex}{An unevaluated R expression, e.g. quote(2 + log(moo - 1))}
}
\value{
A character vector representing the GADGET formulae, e.g. "(+ 2 (log (- #moo 1)))"
}
\description{
Turn R expression into GADGET formulae string
}
