% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{toDataFrame}
\alias{toDataFrame}
\title{toDataFrame}
\usage{
toDataFrame(sim)
}
\arguments{
\item{sim}{}
}
\value{
data.frame
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

