% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGadget.R
\name{variant_append_settings}
\alias{variant_append_settings}
\title{Append setting to variant directories}
\usage{
variant_append_settings(
  gd,
  params.in = attr(gd, "params_in"),
  params.out = NULL,
  control = NULL
)
}
\arguments{
\item{gd}{variant directory}

\item{params.in}{input parameters}

\item{params.out}{output}

\item{control}{optinfo file}
}
\value{
variant dir with appended attributes
}
\description{
More description needed
}
