% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{von_B_formula}
\alias{von_B_formula}
\alias{von_b_formula}
\title{von B formula}
\usage{
von_b_formula(a, linf = "Linf", k = "k", recl = "recl")
}
\arguments{
\item{a}{age}

\item{linf}{value or gadget formula for Linf}

\item{k}{value or gadget formula for 1000*K}

\item{recl}{value or gadget formula mean length at recruit}
}
\value{
vector of von B formula by age
}
\description{
This is a convenience function that creates the von B formulas for the recruitment and initial population files
}
