\name{vonB}
\alias{vonB}
\title{Von Bertalanffy}
\usage{vonB(lsup, k, a)}
\description{Calculate the Von Bertanlaffy curve according to the formula L(a) = L_\infty \big[1-e^{-\kappa a}\big].}
\value{a vector of length(a) with the calculated VB curve at age(s) a.}
\author{Bjarki Þór Elvarsson}
\arguments{\item{lsup}{L_\infty}
\item{k}{\kappa}
\item{a}{age}}
