% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetparameters.R
\name{wide_parameters}
\alias{wide_parameters}
\title{wide parameter files}
\usage{
wide_parameters(dat, value)
}
\arguments{
\item{dat}{gadget parameters object}

\item{value}{data.frame update to the parameter matrix (assumes same row count)}
}
\value{
gadget parameters object
}
\description{
wide parameter files
}
