% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.gadget.area}
\alias{write.gadget.area}
\title{write gadget area file}
\usage{
write.gadget.area(area, file = "area")
}
\arguments{
\item{area}{}

\item{file}{}
}
\value{
nothing
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

