% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{write.gadget.file}
\alias{write.gadget.file}
\title{Write gadgetfile to disk, including any dependant files, and update the mainfile}
\usage{
write.gadget.file(obj, path, mainfile = "main")
}
\arguments{
\item{obj}{gadgetfile object to write}

\item{path}{Base directory to write out to}

\item{mainfile}{The name of the directories mainfile (or NULL to disable mainfile updating)}
}
\description{
Write gadgetfile to disk, including any dependant files, and update the mainfile
}

