% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.gadget.likelihood}
\alias{write.gadget.likelihood}
\title{Write likelihood}
\usage{
write.gadget.likelihood(lik, file = "likelihood", data.folder = NULL,
  bs.sample = NULL)
}
\arguments{
\item{lik}{object of class gadget.likelihood}

\item{file}{name of the likelihood file}

\item{data.folder}{location of data folder (if changed)}

\item{bs.sample}{(for bootstrap), appends the appropriate replicate number to data file}
}
\value{
character string corresponding to the likelihood file (if desired)
}
\description{
Write a likelihood object to file
}
\author{
Bjarki ??r Elvarsson
}

