% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{write.gadget.main}
\alias{write.gadget.main}
\title{Write main}
\usage{
write.gadget.main(main, file = "main")
}
\arguments{
\item{main}{gadget.main object}

\item{file}{name of main file}
}
\value{
text of the main file (if desired)
}
\description{
Write gadget.main object to file
}
\author{
Bjarki ??r Elvarsson
}

