% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.gadget.optinfo}
\alias{write.gadget.optinfo}
\title{Write gadget optinfo}
\usage{
write.gadget.optinfo(optinfo, file = "optinfofile")
}
\arguments{
\item{optinfo}{optinfo object}

\item{file}{file}

\item{location}{location}
}
\value{
text of the optinfofile (if desired)
}
\description{
Write optinfo to file
}
\author{
Bjarki ??r Elvarsson
}

