% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{write.gadget.penalty}
\alias{write.gadget.penalty}
\title{Gadget Penalty file}
\usage{
write.gadget.penalty(file = "penaltyfile")
}
\arguments{
\item{file}{name of the file that is to contain the penalty}
}
\description{
Gadget Penalty file
}

