% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{write.gadget.printfile}
\alias{write.gadget.printfile}
\title{Write Gadget printfile}
\usage{
write.gadget.printfile(print, file = "prinfile", output.dir = "out")
}
\arguments{
\item{print}{printfile object}

\item{file}{string containing the desired location of the printfile}

\item{output.dir}{where should the output from the prinfile components be written, defaults to 'out'.}
}
\value{
(invisible) text of the printfile if desired.
}
\description{
Write the gadget prinfile to file, optionally changing the output directory
of the printfile components.
}
\author{
Bjarki Thor Elvarsson
}
