% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.gadget.time}
\alias{write.gadget.time}
\title{write gadget time file}
\usage{
write.gadget.time(time, file = "time")
}
\arguments{
\item{time}{}

\item{file}{}
}
\value{
nothing
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

