% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{age.length.key}
\alias{age.length.key}
\title{Age length key}
\usage{
age.length.key(sim, age.agg, len.agg)
}
\arguments{
\item{sim}{Results from a Rgadget simulation}

\item{age.agg}{The desired age aggregation}

\item{len.agg}{The desired length aggregation}
}
\value{
Dataframe containing the age - length key
}
\description{
Calculates the age-length-key for the survey and commercial fleet.
}
\author{
Bjarki Þór Elvarsson
}
