% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{as.data.frame.gadget.sim}
\alias{as.data.frame.gadget.sim}
\title{as.data.frame.gadget.sim}
\usage{
\method{as.data.frame}{gadget.sim}(sim)
}
\arguments{
\item{sim}{the results from RGadget}
}
\value{
A dataframe
}
\description{
This function formats the output from RGadget to a dataframe and adds some 
trivial calculated values
}
