\name{as.data.frame.gadget.sim}
\alias{as.data.frame.gadget.sim}
\title{as.data.frame.gadget.sim}
\usage{as.data.frame.gadget.sim(sim)}
\description{This function formats the output from RGadget to a dataframe and adds some 
trivial calculated values}
\value{A dataframe}
\arguments{\item{sim}{the results from RGadget}}
