% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{catch.in.kilos}
\alias{catch.in.kilos}
\title{Catch in Kilos}
\usage{
catch.in.kilos(sim)
}
\arguments{
\item{sim}{Results from a Rgadget simulation}
}
\value{
Dataframe with the catch in kilos by timestep and ared.
}
\description{
Calculates the overall weigth of the catches by time step and area.
}
\author{
Bjarki Þór Elvarsson
}
