\name{catch.in.kilos}
\alias{catch.in.kilos}
\title{Catch in Kilos}
\usage{catch.in.kilos(sim)}
\description{Calculates the overall weigth of the catches by time step and area.}
\value{Dataframe with the catch in kilos by timestep and ared.}
\author{Bjarki Þór Elvarsson}
\arguments{\item{sim}{Results from a Rgadget simulation}}
