% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{distr}
\alias{distr}
\title{Length distribution}
\usage{
distr(mu, sigma, l, par = data.frame())
}
\arguments{
\item{mu}{mean length for all ages}

\item{sigma}{standard deviation of length for all ages}

\item{l}{lengthgroups}

\item{par}{gadget parameters objects}
}
\value{
a matrix of dimension length(mu) X (length(l)-1)
}
\description{
This is a helper function for the firststep function. This defines the
length distribution for each age group
}
