\name{distr}
\alias{distr}
\title{Length distribution}
\usage{distr(mu, sigma, l)}
\description{This is a helper function for the firststep function. This defines the
length distribution for each age group}
\value{a matrix of dimension length(mu) X (length(l)-1)}
\arguments{\item{mu}{mean length for all ages}
\item{sigma}{standart deviation of length for all ages}
\item{l}{lengthgroups}}
