% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{gadget-growth-class}
\alias{gadget-growth-class}
\title{Gadget growth}
\description{
Gadget growth
}
\section{Slots}{

\describe{
\item{\code{growthfunction}}{}

\item{\code{growthimplementation}}{}

\item{\code{growthparameters}}{}

\item{\code{wgrowthparameters}}{}

\item{\code{lgrowthparameters}}{}

\item{\code{weightgrowthdata}}{}

\item{\code{yeareffect}}{}

\item{\code{stepeffect}}{}

\item{\code{areaeffect}}{}

\item{\code{beta}}{}

\item{\code{maxlengthgroupgrowth}}{}
}}

