% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetClass.R
\docType{class}
\name{gadget-growth-class}
\alias{gadget-growth-class}
\title{Gadget growth}
\description{
Gadget growth
}
\section{Slots}{

\describe{
\item{\code{growthfunction}}{name of the growth function}

\item{\code{growthimplementation}}{name of the growth implementation function}

\item{\code{growthparameters}}{parameters for the growth function}

\item{\code{wgrowthparameters}}{weight related growth function parameters}

\item{\code{lgrowthparameters}}{length related growth function parameters}

\item{\code{yeareffect}}{year effect}

\item{\code{stepeffect}}{step effect}

\item{\code{areaeffect}}{area effect}

\item{\code{beta}}{beta for the growth implementation}

\item{\code{weigthgrowthdata}}{data on growth in weight}

\item{\code{maxlengthgroupgrowth}}{maximum length group growth}
}}

