% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.forward.R
\name{gadget.bootforward}
\alias{gadget.bootforward}
\title{Gadget bootstrap forward}
\usage{
gadget.bootforward(years = 20, params.file = "params.final",
  main.file = "main.final", pre = "PRE", effort = 0.2,
  fleets = data.frame(fleet = "comm", ratio = 1), num.trials = 10,
  bs.wgts = "BS.WGTS", bs.samples = 1:1000, check.previous = FALSE,
  rec.window = NULL, mat.par = NULL, stochastic = TRUE,
  .parallel = TRUE)
}
\value{
list of bootstrap results
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}
