% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfit.R
\name{gadget.fit}
\alias{gadget.fit}
\title{Gadget fit}
\usage{
gadget.fit(wgts = "WGTS", main.file = NULL, fleet.predict = NULL,
  mat.par = NULL, params.file = NULL, f.age.range = NULL,
  fit.folder = "FIT", printfile.printatstart = 1, printfile.steps = 1,
  rec.length.param = NULL)
}
\arguments{
\item{wgts}{Location of the iterative reweighting folder, if null gadget.fit requires a parameter file}

\item{main.file}{if the main file is different from the 'main' or 'wgts/main.final'}

\item{fleet.predict}{data.frame containing the basis fleets used to calculate the harvestable biomass. Fleet names should be specified in the fleet column.}

\item{mat.par}{parameters for the maturity ogive}

\item{params.file}{parameter file used in the fit (defaults to "WGTS/params.final")}

\item{f.age.range}{data.frame describing the desired age range where the F's are calculated, if null this defaults to the apical F fro all stocks. 
Input columns should include stock, age.min and age.max}

\item{fit.folder}{location of the output}

\item{printfile.printatstart}{should the stock standared output be printed at the beginning or the end of the timestep}

\item{printfile.steps}{what steps should be printed}

\item{rec.length.param}{Character. The parameter that determines recruitment length in rec files (if used)}
}
\value{
list of fit things
}
\description{
This function calls Gadget and compiles fit statistics, relevant biomass trajectories and consumption
}
\author{
Bjarki Thor Elvarsson
}
