% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{gadget.fit}
\alias{gadget.fit}
\title{Gadget fit}
\usage{
gadget.fit(wgts = "WGTS", main.file = "main",
  fleet.predict = data.frame(fleet = "comm", ratio = 1), mat.par = NULL,
  params.file = NULL, f.age.range = NULL, fit.folder = "FIT")
}
\value{
list of fit things
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

