% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfit.R
\name{gadget.fit}
\alias{gadget.fit}
\alias{bind.gadget.fit}
\title{Gadget fit}
\usage{
gadget.fit(wgts = "WGTS", main.file = NULL, fleet.predict = NULL,
  mat.par = NULL, params.file = NULL, f.age.range = NULL,
  fit.folder = "FIT", printfile.printatstart = 1,
  printfile.steps = 1, rec.len.param = FALSE)

bind.gadget.fit(...)
}
\arguments{
\item{wgts}{Location of the iterative reweighting folder, if null gadget.fit requires a parameter file}

\item{main.file}{if the main file is different from the 'main' or 'wgts/main.final'}

\item{fleet.predict}{data.frame containing the basis fleets used to calculate the harvestable biomass. Fleet names should be specified in the fleet column.}

\item{mat.par}{parameters for the maturity ogive}

\item{params.file}{parameter file used in the fit (defaults to "WGTS/params.final")}

\item{f.age.range}{data.frame describing the desired age range where the F's are calculated, if null this defaults to the apical F fro all stocks. 
Input columns should include stock, age.min and age.max}

\item{fit.folder}{location of the output}

\item{printfile.printatstart}{should the stock standared output be printed at the beginning or the end of the timestep}

\item{printfile.steps}{what steps should be printed}

\item{rec.len.param}{Logical. TRUE if you want growth calculated as age.based and using a formula other than get.gadget.growth default}

\item{...}{any number of gadget.fit objects}
}
\value{
list containing the output from Gadget.
}
\description{
\code{gadget.fit} calls Gadget and compiles fit statistics, relevant biomass trajectories and consumption.
It returns an object of class 'gadget.fit' which is essentially a list containing the following objects: \describe{
\item{sidat}{data.frame containing the model fit to the surveyindices}
\item{resTable}{Table of likelihood component scores from the different stages of the iterative reweighting run}
\item{nesTable}{Same as resTable but normalised with the minimum score for each component}
\item{suitability}{The model estimated fleet predator suitability}
\item{stock.growth}{Model estimated growth}
\item{stock.recruitment}{Recruitment by stock}
\item{res.by.year}{Results by year}
\item{stomachconten}{Fit combined with data from the stomachcontent likelihood}
\item{likelihoodsummary}{Likelihood component scores by year and step} 
\item{catchdist.fleets}{Data compared with model output for the catchdistribution components}
\item{stockdist}{Data compared with model output for the stockdistribution components}
\item{SS}{The likelihood output from the model}
\item{stock.full}{Modeled abundance and mean weight by year,step, length and stock}
\item{stock.std}{Modeled abundance, mean weight, number consumed by age, stock and year}
\item{stock.prey}{Consumption of prey (or fleet catches) by prey, length, year and step}
\item{fleet.info}{Information on catches, harvest rate and harvestable biomass by fleet, year and step}
\item{predator.prey}{consumption of preys by predators, length, year and step}
\item{params}{parameter values used for the fit}
\item{catchstatistics}{Data compared with model output for the catchdstatistics components}
}

\code{bind.gadget.fit} combines two (or more) gadget.fit objects into a 
single list. Each object in the list will have an extra column indicating the model, 
if the list is named those names will be used to identify the model.
}
\author{
Bjarki Thor Elvarsson
}
