% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{gadget.forward}
\alias{gadget.forward}
\title{Gadget forward simulation}
\usage{
gadget.forward(years = 20, params.file = "params.out", main.file = "main",
  num.trials = 10, fleets = data.frame(fleet = "comm", ratio = 1),
  biomass = FALSE, effort = 0.2, spawnmodel = "none", spawnvar = NULL,
  selectedstocks = NULL, biomasslevel = NULL, check.previous = FALSE,
  save.results = TRUE, stochastic = TRUE, rec.window = NULL,
  compact = TRUE, mat.par = c(0, 0), gd = list(dir = ".", rel.dir =
  "PRE"))
}
\value{
list of results
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

