% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetoptions.R
\name{gadget.options}
\alias{gadget.options}
\title{Gagdet options}
\usage{
gadget.options(type = c("simple2stock", "spawning"))
}
\arguments{
\item{type}{}
}
\value{
a list of swithes
}
\description{
This function creates a list of default values of all necessary
switches for gadget.simulation and gadget.skeleton functions. The
user can then change the values of the switches and use the
changed list as input to those functions.  Here the default values for
consumption (predation and fleet operations), migration,
maturation (via stock 'movement'), number of areas and their basic
properties, various attributes of the sub stocks such as age,
length and weight along with growth parameters. Also length of the
simulation is given a default value.  If the user wants to change
the default values he will need to make the changes on the
resulting list.
}
\examples{
opt <- gadget.options
## change the length of the simulation to 13 years
opt$numobs <- 13
}
\author{
Bjarki Thor Elvarsson
}
