% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.parameters}
\alias{read.gadget.parameters}
\alias{write.gadget.parameters}
\title{Gadget parameters}
\usage{
read.gadget.parameters(file = "params.in")

write.gadget.parameters(params, file = "params.out", columns = TRUE)
}
\arguments{
\item{file}{parameter file}

\item{params}{params dataframe}

\item{columns}{should a conventional column based parameter file be written out or
should a line based parameter (used when evaluating gadget on a matrix) be used.}

\item{file}{a string naming the file to write to}
}
\value{
dataframe

a string containing the text of the params file (if desired)
}
\description{
\code{read.gadget.parameters} reads a gadget parameter file
}
\author{
Bjarki  Thor Elvarsson

Bjarki ??r Elvarsson
}
