% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{gadget.phasing}
\alias{gadget.phasing}
\title{Gadget Phasing}
\usage{
gadget.phasing(phase, params.in = "params.in", main = "main",
  phase.dir = "PHASING", optinfofile = "optinfofile")
}
\arguments{
\item{phase}{a dataframe where the columns indicate the parameters
that are to be optimised in that particular phase}

\item{params.in}{either a filename or gadget.parameters object
containing the initial value for the optimisation.}

\item{main}{name of the main file used in the optimisation.}

\item{phase.dir}{output directory}

\item{optinfofile}{location of the optinfofile}
}
\value{
final optimised parameter values
}
\description{
Phased minimization based on variables
}
\author{
Bjarki Thor Elvarsson
}
