% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{gadget.retro}
\alias{gadget.retro}
\title{gadget.retro}
\usage{
gadget.retro(path = ".", main.file = "main", params.file = "params.in",
  optinfo.file = "optinfofile", num.years = 5, pre = "RETRO")
}
\arguments{
\item{path}{}

\item{main.file}{}

\item{params.file}{}

\item{optinfo.file}{}

\item{num.years}{}

\item{pre}{}
}
\description{
gadget.retro
}

