% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfunctions.R
\name{gadget.ypr}
\alias{gadget.ypr}
\alias{plot.gadget.ypr}
\title{Gadget Yield per Recruit}
\usage{
gadget.ypr(params.file = "params.in", main.file = "main",
  effort = seq(0, 1, by = 0.01), begin = 1990, end = 2020,
  age.range = NULL, fleets = data.frame(fleet = "comm", ratio = 1),
  ssb.stock = NULL, mat.par = NULL, check.previous = FALSE,
  save.results = TRUE, gd = list(dir = ".", rel.dir = "YPR"))

\method{plot}{gadget.ypr}(ypr)
}
\arguments{
\item{params.file}{Parameter file for the gagdet model}

\item{main.file}{Main file for the gagdet model}

\item{effort}{The range of fishing mortality}

\item{begin}{Start year of the simulation}

\item{end}{End year of the simulation}

\item{age.range}{at what age range should the YPR be calculated}

\item{fleets}{Data frame comtaining the fleet names and ratio in
future catches}

\item{check.previous}{check if the analysis have been done before}

\item{save.results}{should the results be saved?}

\item{gd}{gadget directory object}
}
\value{
a list containing the yield per recruit by F, estimate of
Fmax and F0.1
}
\description{
Calculate yield per recruit of a stock in a Gadget model
}
\details{
\code{gadget.ypr} returns the results from a yield per recruit simulation. 
The function sets up a new Gadget run based on a given model 
and parameter settings where the initial conditions and recruiment 
parameters are all set to zero appart from one year class which is 
set to a nominal value (10000 individuals). The model is the run with 
a range of fishing effort settings. The fishing effort is formulated in
relation to linearfleets whose selection is the same as the one in the 
fleets in the model. The user can specify what fleets will be used in
in the yield per recruit simulation.
}
\examples{
\dontrun{
ypr <- gadget.ypr(ssb.stock = 'codmat',
                  params.file = 'WGTS/params.final')
plot(ypr)
}
}
\author{
Bjarki Thor Elvarsson
}
