% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_catchdistribution_component}
\alias{gadget_catchdistribution_component}
\title{http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:catchdist}
\usage{
gadget_catchdistribution_component(
  weight = 0,
  name = "catchdistribution",
  data_function = "sumofsquares",
  data_function_params = list(),
  aggregationlevel = FALSE,
  overconsumption = FALSE,
  epsilon = 10,
  data = NULL,
  area = NULL,
  age = NULL,
  length = NULL,
  fleetnames = c(),
  stocknames = c()
)
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}

\item{data_function}{Function to be used to compare the modelled statistical data to the input statistical data}

\item{data_function_params}{Parameters if using a multivariate function}

\item{aggregationlevel}{Boolean. aggregate data over whole year? Default false.}

\item{overconsumption}{Boolean, take overconsumption into account? Default false.}

\item{epsilon}{Used when probability is very unikely, default 10.}

\item{data}{data.frame required by the data_function}

\item{area}{Area aggregation, if not attached to the data.frame by MFDB}

\item{age}{Age aggregation, if not attached to the data.frame by MFDB}

\item{length}{Length aggregation, if not attached to the data.frame by MFDB}

\item{fleetnames}{A list of all the fleets to be aggregated into a single pseudo fleet for the purposes of the data comparison}

\item{stocknames}{A list of all the stocks to be aggregated into a single pseudo stock}
}
\description{
http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:catchdist
}
