% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_fit.R
\name{gadget_fit}
\alias{gadget_fit}
\title{Gadget fit}
\usage{
gadget_fit(
  gd,
  params.in = attr(gd, "params_in"),
  fit = "FIT",
  f.age.range = NULL,
  weighted.mortality = FALSE,
  rec.age = NULL,
  rec.steps = NULL,
  steps = 1
)
}
\arguments{
\item{gd}{gadget.variant.dir}

\item{params.in}{input parameters}

\item{fit}{location of the fit folder}

\item{f.age.range}{age range for F}

\item{weighted.mortality}{Method for calculating annual mortality. TRUE for a weighted mean, FALSE for an unweighted mean. This is only applicable if f.age.range is specified. Defaults to FALSE.}

\item{rec.steps}{Step(s) at which recruitment occurs, defaults to "all" & applies to all stocks. Note that recruitment is printed at the start of the timestep.}

\item{steps}{(unused)}

\item{rec.at.age}{Age at which recruitment occurs, integer value that applies across stocks.}
}
\value{
list
}
\description{
Gadget fit
}
