% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_migrationpenalty_component}
\alias{gadget_migrationpenalty_component}
\title{http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:migpenalty}
\usage{
gadget_migrationpenalty_component(weight = 0,
  name = "migrationpenalty", stockname = c(), powercoeffs = c())
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:migpenalty
}
