% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_penalty_component}
\alias{gadget_penalty_component}
\title{https://hafro.github.io/gadget/docs/userguide#boundlikelihood_(%E2%80%9Cpenalty%E2%80%9D)}
\usage{
gadget_penalty_component(weight = 0, name = "penalty", data = NULL)
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
https://hafro.github.io/gadget/docs/userguide#boundlikelihood_(%E2%80%9Cpenalty%E2%80%9D)
}
