% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_predatoroverprinter_component}
\alias{gadget_predatoroverprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:predatoroverprinter}
\usage{
gadget_predatoroverprinter_component(
  printfile,
  predatornames,
  area,
  len,
  precision = 15,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{Name for the output file to be created}

\item{predatornames}{Names of the predators}

\item{area}{list of area aggregations    agere}

\item{len}{list of length aggregations}

\item{precision}{optional value defining numerical precision}

\item{yearsandsteps}{character defining when to print, defaults to "all all"}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:predatoroverprinter
}
