% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_predatorprinter_component}
\alias{gadget_predatorprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:predatorprinter}
\usage{
gadget_predatorprinter_component(
  printfile,
  predatornames,
  preynames,
  area,
  predlen,
  preylen,
  biomass = 1,
  precision = 15,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{Name for the output file to be created}

\item{predatornames}{Names of the predators}

\item{preynames}{Names of the prey stocks}

\item{area}{list of area aggregations    agere}

\item{predlen}{list of length aggregations}

\item{preylen}{list of length aggregations}

\item{biomass}{should biomass (1) or abundance (0) be printed}

\item{precision}{optional value defining numerical precision}

\item{yearsandsteps}{character defining when to print, defaults to "all all"}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:predatorprinter
}
