% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetprintcomponent.R
\name{gadget_stockstdprinter_component}
\alias{gadget_stockstdprinter_component}
\title{https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockstdprinter}
\usage{
gadget_stockstdprinter_component(
  printfile,
  stockname,
  scale = 1,
  precision = 15,
  printatstart = 1,
  yearsandsteps = "all\\tall"
)
}
\arguments{
\item{printfile}{Name for the output file to be created}

\item{stockname}{Name of the stock}

\item{scale}{The datafile lists these weights and the power that is to be used for each parameter}

\item{precision}{optional value defining numerical precision}

\item{printatstart}{should the output be printed at the start (1) or the end (0) of the timesstep}

\item{yearsandsteps}{character defining when to print, defaults to "all all"}
}
\description{
https://hafro.github.io/gadget2/userguide/chap-print.html#sec:stockstdprinter
}
