% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadget_stomachcontent_component}
\alias{gadget_stomachcontent_component}
\title{http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:stomach}
\usage{
gadget_stomachcontent_component(weight = 0, name = "stomachcontent",
  data_function = "scsimple", epsilon = 10, area = NULL,
  predator_length = NULL, prey_length = NULL, prey_labels = list(),
  prey_digestion_coefficients = list(), predator_names = c(),
  data = NULL)
}
\arguments{
\item{weight}{A numeric weighting}

\item{name}{A descriptive name for the component, default same as component type}
}
\description{
http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:stomach
}
