% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfleet.R
\name{gadget_update.gadgetfleet}
\alias{gadget_update.gadgetfleet}
\title{Update gadget fleet components in a fleet file}
\usage{
\method{gadget_update}{gadgetfleet}(gf, component, ...)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{component}{Either a replacement \code{gadget_fleet_component} (from MFDB or rgadget), or a component type name}

\item{...}{If a component type was provided above, the extra options to supply to \code{gadget_fleet_component}}
}
\description{
Replace and/or append new fleet comonents to an existing file
}
\examples{
path <- './model'
gadgetfleet('fleet', path, missingOkay = TRUE) \%>\%
   gadget_update(gadget_fleet_component('understocking', name = 'understocking')) \%>\% # Add an understocking component
   gadget_update('understocking', name = 'understocking') \%>\% # Identical to above
   write.gadget.file(path)
}
