% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgettime.R
\name{gadget_update.gadgettime}
\alias{gadget_update.gadgettime}
\title{Update gadget time components in a time file}
\usage{
\method{gadget_update}{gadgettime}(gf, ...)
}
\arguments{
\item{gf}{The gadgetfile object to update}

\item{...}{If a component type was provided above, the extra options to supply to \code{gadget_time_component}}

\item{component}{Either a replacement \code{gadget_time_component} (from MFDB or rgadget), or a component type name}
}
\description{
Replace and/or append new time comonents to an existing file
}
\examples{
path <- './model'
gadgettime('time', path, missingOkay = TRUE) \%>\%
   gadget_update() \%>\% 
   write.gadget.file(path)
}
