% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.likelihood}
\alias{read.gadget.likelihood}
\alias{write.gadget.likelihood}
\alias{merge.gadget.likelihood}
\alias{get.gadget.likelihood}
\alias{read.gadget.main}
\alias{write.gadget.main}
\alias{clear.spaces}
\alias{read.gadget.data}
\alias{strip.comments}
\alias{read.gadget.stockfiles}
\alias{read.gadget.area}
\alias{write.gadget.area}
\alias{read.gadget.time}
\alias{write.gadget.time}
\alias{read.gadget.table}
\alias{read.gadget.fleet}
\alias{write.gadget.fleet}
\title{Old style gadget file input and output (mostly deprecated)}
\usage{
read.gadget.likelihood(files = "likelihood")

write.gadget.likelihood(lik, file = "likelihood", data.folder = NULL,
  bs.sample = NULL)

\method{merge}{gadget.likelihood}(lik1, lik2)

get.gadget.likelihood(likelihood, comp, inverse = FALSE)

read.gadget.main(file = "main")

write.gadget.main(main, file = "main")

clear.spaces(text)

read.gadget.data(likelihood, debug = FALSE, year_range = NULL)

strip.comments(file = "main")

read.gadget.stockfiles(stock.files)

read.gadget.area(area.file = "area")

write.gadget.area(area, file = "area")

read.gadget.time(time.file = "time")

write.gadget.time(time, file = "time")

read.gadget.table(file, header = FALSE)

read.gadget.fleet(fleet.file = "fleet")

write.gadget.fleet(fleet, file = "fleet")
}
\arguments{
\item{files}{a vector of character strings containing the names of the likelihood files}

\item{lik}{object of class gadget.likelihood}

\item{file}{name of the likelihood file}

\item{data.folder}{location of data folder (if changed)}

\item{bs.sample}{(for bootstrap), appends the appropriate replicate number to data file}

\item{lik1}{likelihood object}

\item{lik2}{likelihood object}

\item{likelihood}{likelihood object}

\item{comp}{selected likelihood components}

\item{inverse}{(logical) should inverse selection be applied}

\item{main}{gadget.main object}

\item{text}{string}

\item{debug}{should debug information be printed}

\item{stock.files}{location of stock file}

\item{area.file}{location of the area file}

\item{area}{data frame with area and temperature}

\item{time.file}{location of the file}

\item{time}{data.frame with time}

\item{header}{logical, should the header be read from the file}

\item{fleet.file}{location of the fleet file}

\item{fleet}{a gadget.fleet}

\item{file}{main file location}

\item{file}{name of main file}

\item{likelihood}{object of class gadget.likelihood}

\item{file}{location of the gadget input file}

\item{file}{location of the area file}

\item{file}{location of the time file}

\item{file}{path to file}

\item{file}{location of the file}
}
\value{
object of class gadget.likelihood, i.e. a list containing the various likelihood components

character string corresponding to the likelihood file (if desired)

merged gadget likelihood object

likelihood object

object of class gadget.main

text of the main file (if desired)

list or matrix containing the (non-empty) values from the string

list of dataframes and degress of freedom

list containing the lines from the file stripped of unwanted text.

list of gadget-stock objects

areafile object

nothing

timefile objects

nothing

data.frame

fleet file object

nothing
}
\description{
\code{read.gadget.likelihood} reads the likelihood (input) file for gadget. The format of
the likelihood file is described in gadget's user manual.

\code{read.gadget.data} reads data used by the various components

\code{read.gadget.stockfiles} reads in Gadget stockfiles DEPRECATED

\code{read.gadget.area} reads the gadget area file

\code{write.gadget.area} writes a gadget area file

\code{read.gadget.fleet} read gadget fleet
}
\author{
Bjarki ??r Elvarsson

Bjarki ??r Elvarsson

Bjarki Thor Elvarsson

Bjarki Thor Elvarsson

Bjarki ??r Elvarsson

Bjarki ??r Elvarsson

Bjarki ??r Elvarsson

Bjarki ??r Elvarsson

Bjarki Thor Elvarsson

Bjarki Thor Elvarsson

Bjarki Thor Elvarsson
}
