% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfleetcomponent.R
\name{gadgetfleetcomponent}
\alias{gadgetfleetcomponent}
\title{Create/update a fleet component with name}
\usage{
gadgetfleetcomponent(type, name = type,
  livesonareas = unique(data$area), multiplicative = 1,
  suitability = NULL, fleetfile = "fleet",
  data = stop("data not provided"), ...)
}
\description{
Create/update a fleet component with name
}
