% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetlikelihoodcomponent.R
\name{gadgetlikelihoodcomponent}
\alias{gadgetlikelihoodcomponent}
\title{Wrapper to choose a component by name}
\usage{
gadgetlikelihoodcomponent(component, ...)
}
\arguments{
\item{component}{Call the corresponding gadget_(component)_component function}

\item{...}{Arguments to pass to it}
}
\description{
Wrapper to choose a component by name
}
