% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{get.gadget.catches}
\alias{get.gadget.catches}
\title{get gadget catches}
\usage{
get.gadget.catches(fleets, params)
}
\value{
catches by year + step
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

