% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.gadget.growth}
\alias{get.gadget.growth}
\title{get gadget growth}
\usage{
get.gadget.growth(stocks, params, dt = 0.25, age.based = FALSE)
}
\arguments{
\item{stocks}{}

\item{params}{}

\item{dt}{}

\item{age.based}{}
}
\value{
growth matrix
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

