% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{get.gadget.growth}
\alias{get.gadget.growth}
\title{get gadget growth}
\usage{
get.gadget.growth(stocks, params, dt = 0.25, age.based = FALSE)
}
\value{
growth matrix
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}
