% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{get.gadget.recruitment}
\alias{get.gadget.recruitment}
\title{get gadget recruitment}
\usage{
get.gadget.recruitment(stocks, params, collapse = TRUE)
}
\arguments{
\item{stocks}{gadget.stock}

\item{params}{gadget.params}
}
\value{
recruitment by year
}
\author{
Bjarki Thor Elvarsson
}
