% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.gadget.recruitment}
\alias{get.gadget.recruitment}
\title{get gadget recruitment}
\usage{
get.gadget.recruitment(stocks, params)
}
\arguments{
\item{stocks}{}

\item{params}{}
}
\value{
recruitment by year
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

