% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.gadget.suitability}
\alias{get.gadget.suitability}
\title{get gadget suitability}
\usage{
get.gadget.suitability(fleets, params, lengths, normalize = FALSE)
}
\arguments{
\item{fleets}{}

\item{params}{}

\item{lengths}{}

\item{normalize}{}
}
\value{
suitability function for a fleet
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

