\name{lengthDist}
\alias{lengthDist}
\title{Length Dist}
\usage{lengthDist(sim, sigma=0)}
\description{Calculate the length distribution from the fleet by length groups and time.}
\value{Dataframe containing the fleet length distribution.}
\author{Bjarki Þór Elvarsson}
\arguments{\item{sim}{Results from a Rgadget simulation}
\item{sigma}{sigma for a lognormal noise}}
