% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{make.gadget.printfile}
\alias{make.gadget.printfile}
\title{Make gadget printfile}
\usage{
make.gadget.printfile(main = "main", output = "out",
  aggfiles = "print.aggfiles", file = "printfile")
}
\arguments{
\item{main}{location of main file}

\item{output}{location of output folder}

\item{aggfiles}{location of aggfiles}

\item{file}{name of resulting printfile}
}
\value{
gadget.mainfile object
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

