% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{merge.formula}
\alias{merge.formula}
\title{merge txt to gadget.formula}
\usage{
\method{merge}{formula}(txt)
}
\value{
txt where the formula has been properly compiled
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}
