% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{migrate}
\alias{migrate}
\title{Migrate}
\usage{
migrate(N, M)
}
\arguments{
\item{N}{An array containing substock abundance by area (first dimension) and other variables}

\item{M}{an array with migration matricies}
}
\value{
the migrated substock abundance array
}
\description{
If more than one area is defined the user can define migration
between the areas. The migration procedure is the same as in MULTSPEC. 
The migration is defined and takes place in every timestep and it is
assumed to be constant for all years. This means that the porportion that
migrates from area i to j can be different between different timesteps
(seasons) but they are fixed between observations (years). Migration at
timestep t is defined by the user as an \eqn{n\times n}{nxn} transition
matrix P_t := [p_{ij}] where p_{ij} is the proportion moving from area j
to area i, ignoring other processes. For P to be a transition matrix
\eqn{\sum_ip_{ij} = 1}{sum_i p_ij = 1}, for each j. The vector of abundance
for all areas at time t is therefore:
\deqn{\mathbf{N_t} = P_t \mathbf{N_{t-1}}}{N_t = P_t N_{t-1}}
In a two area example this would mean that if N_{1,t-1} is a matrix
containing the abundance in numbers in area 1 at timestep t before
migration and N_{2,t-1} is the same number for area 2, the numbers after
migration will be
\deqn{N_{1,t} = p_{11}\cdot N_{1,t-1} + p_{12}\cdot N_{2,t-1} }
\deqn{N_{2,t} = p_{21}\cdot N_{2,t-1}+ p_{22}\cdot N_{2,t-1} }
}
\author{
Bjarki Þór Elvarsson
}
