% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{plot.gadget.options}
\alias{plot.gadget.options}
\title{Plot gadget options}
\usage{
\method{plot}{gadget.options}(opt)
}
\arguments{
\item{opt}{Gadget options object}
}
\description{
Plots the suitability functions, the length weight relationship and the Von Bertalanffy growth curve.
}
\author{
Bjarki Þór Elvarsson
}
