\name{plot.gadget.options}
\alias{plot.gadget.options}
\title{Plot gadget options}
\usage{plot.gadget.options(opt)}
\description{Plots the suitability functions, the length weight relationship and the Von Bertalanffy growth curve.}
\author{Bjarki Þór Elvarsson}
\arguments{\item{opt}{Gadget options object}}
