% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{plot.Rgadget}
\alias{plot.Rgadget}
\title{Plot Rgadget}
\usage{
\method{plot}{Rgadget}(dat, compare.alk = TRUE)
}
\arguments{
\item{dat}{A Rgadget object}
}
\description{
Plot the results from the summary functions of the Rgadget simulation.
}
\author{
Bjarki Þór Elvarsson
}
