% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.area}
\alias{read.gadget.area}
\title{read gadget area file}
\usage{
read.gadget.area(area.file = "area")
}
\arguments{
\item{area.file}{}
}
\value{
areafile object
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

