% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.data}
\alias{read.gadget.data}
\title{Read likelihood data}
\usage{
read.gadget.data(likelihood, debug = FALSE, year_range = NULL)
}
\arguments{
\item{likelihood}{object of class gadget.likelihood}

\item{debug}{should debug information be printed}
}
\value{
list of dataframes and degress of freedom
}
\description{
Read data used by the various components
}
\author{
Bjarki ??r Elvarsson
}

