% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.data}
\alias{read.gadget.data}
\title{Read likelihood data}
\usage{
read.gadget.data(likelihood)
}
\arguments{
\item{likelihood}{object of class gadget.likelihood}
}
\value{
list of dataframes and degress of freedom
}
\description{
Read data used by the various components
}
\author{
Bjarki ??r Elvarsson
}

