% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gadget.fleet}
\alias{read.gadget.fleet}
\title{read gadget fleet}
\usage{
read.gadget.fleet(fleet.file = "fleet")
}
\arguments{
\item{fleet.file}{}
}
\value{
fleet file object
}
\description{
<description>
}
\details{
function only tested for linear and totalfleets>
}
\author{
Bjarki Þór Elvarsson
}

