% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.fleet}
\alias{read.gadget.fleet}
\title{read gadget fleet}
\usage{
read.gadget.fleet(fleet.file = "fleet")
}
\value{
fleet file object
}
\description{
<description>
}
\details{
function only tested for linear and totalfleets>
}
\author{
Bjarki Þór Elvarsson
}

