% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.grouping}
\alias{read.gadget.grouping}
\title{read gadget grouping}
\usage{
read.gadget.grouping(lik = read.gadget.likelihood(), wgts = "WGTS")
}
\arguments{
\item{lik}{}

\item{wgts}{}
}
\value{
list of wgts groupings
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

