% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.grouping}
\alias{read.gadget.grouping}
\title{read gadget grouping}
\usage{
read.gadget.grouping(lik = read.gadget.likelihood(), wgts = "WGTS")
}
\value{
list of wgts groupings
}
\author{
Bjarki Thor Elvarsson
}
