% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.model}
\alias{read.gadget.model}
\title{read gadget main}
\usage{
read.gadget.model(main.file = "main", model.name = "Gadget-model")
}
\arguments{
\item{main.file}{location of the main file}
}
\value{
mainfile object
}
\description{
Read in a Gadget model to a gadget-main object DEPRECATED
}
\author{
Bjarki Thor Elvarsson
}
