% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.parameters}
\alias{read.gadget.parameters}
\title{Read param}
\usage{
read.gadget.parameters(file = "params.in")
}
\arguments{
\item{file}{parameter file}
}
\value{
dataframe
}
\description{
Read gadget parameter file
}
\author{
Bjarki  Thor Elvarsson
}

