% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gadget.stockfiles}
\alias{read.gadget.stockfiles}
\title{read gadget stock}
\usage{
read.gadget.stockfiles(stock.files)
}
\arguments{
\item{stock.files}{}
}
\value{
list of gadget-stock objects
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

