% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.table}
\alias{read.gadget.table}
\title{read.gadget.table}
\usage{
read.gadget.table(file, header = FALSE)
}
\arguments{
\item{file}{path to file}

\item{header}{logical, should the header be read from the file}
}
\value{
data.frame with
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

