% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.time}
\alias{read.gadget.time}
\title{read gadget time file}
\usage{
read.gadget.time(time.file = "time")
}
\value{
timefile objects
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}

