% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{read.gadget.wgtsprint}
\alias{read.gadget.wgtsprint}
\title{read wgts print}
\usage{
read.gadget.wgtsprint(wgts = "WGTS", comp = "final", out.pre = "out.",
  parallel = FALSE)
}
\value{
printfiles from the wgts
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}
