% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gadget.wgtsprint}
\alias{read.gadget.wgtsprint}
\title{read wgts print}
\usage{
read.gadget.wgtsprint(wgts = "WGTS", comp = "final", out.pre = "out.",
  parallel = FALSE)
}
\arguments{
\item{wgts}{}

\item{comp}{}

\item{out.pre}{}

\item{parallel}{}
}
\value{
printfiles from the wgts
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Bjarki Thor Elvarsson
}

