% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{suitability}
\alias{suitability}
\title{Prey suitability}
\usage{
suitability(params, l, L = c(0), type = "exponential",
  to.data.frame = FALSE, normalize = FALSE)
}
\arguments{
\item{l}{prey length group(s)}

\item{L}{predator length group(s)}

\item{salpha}{\eqn{\alpha}{alpha} for the suitability function.}

\item{sbeta}{\eqn{\beta}{beta} for the suitability function.}

\item{sgamma}{\eqn{\gamma}{gamma} for the suitability function.}

\item{sdelta}{\eqn{\delta}{delta} for the suitability function.}
}
\value{
matrix of suitabilities, columns prey length, lines predator length
}
\description{
The suitability function for predation used in the R model is:
\deqn{S_{pred,prey}(L,l) = \frac{\delta}{1+e^{-\alpha-\beta l-\gamma L}}}{S_{pred,prey}(L,l) = \frac{delta}{1+e^{-alpha-beta l-gamma L}}}
With one predator, one prey and otherfood the equation becomes:
\deqn{C_{L,l}=N_{L}M_{L}\Psi_{L}\frac{F_{L,l}}{\sum_lF_{L,l}+OA}}
\deqn{=N_{L}M_{L}\frac{F_{L,l}}{\sum_lF_{L,l}+OA+HA}}
where $O$ is the density of otherfood.
}
