% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{summary.gadget.options}
\alias{summary.gadget.options}
\title{Summary of gadget.options}
\usage{
\method{summary}{gadget.options}(opt)
}
\arguments{
\item{opt}{gadget.options list}
}
\description{
summary of the simulation defined by gadget.options
}
