% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{survey.index}
\alias{survey.index}
\title{Survey indices}
\arguments{
\item{sim}{Results from a Rgadget simulation}

\item{sigma}{sigma for a log-normal noise for the indicies}
}
\value{
Dataframe with the survey indices
}
\description{
Calculates the survey indices for the simulated stocks.
}
\author{
Bjarki Þór Elvarsson
}
