% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{survey.index}
\alias{survey.index}
\title{Survey indices}
\usage{
survey.index(stock.dat, split, sigma = 0, alpha = 0, beta = 1)
}
\arguments{
\item{stock.dat}{Results from a Rgadget simulation}

\item{split}{Number of length groups to divide data into}

\item{sigma}{sigma for a log-normal noise for the indicies}

\item{alpha}{log(q), i.e. the constant term}

\item{beta}{power for the N, i.e. the slope term}
}
\value{
Dataframe with the survey indices
}
\description{
Calculates the survey indices for the simulated stocks.
}
\author{
Bjarki Þór Elvarsson
}
