\name{survey.index}
\alias{survey.index}
\title{Survey indices}
\usage{survey.index(sim, sigma=0)}
\description{Calculates the survey indices for the simulated stocks.}
\value{Dataframe with the survey indices}
\author{Bjarki Þór Elvarsson}
\arguments{\item{sim}{Results from a Rgadget simulation}
\item{sigma}{sigma for a log-normal noise for the indicies}}
