\name{survey.indexlen}
\alias{survey.indexlen}
\title{Survey length index}
\usage{survey.indexlen(sim, length.groups=c(4, 14, 90), sigma=0)}
\description{Calculate the survey length index based on the provided lengthgroups}
\value{Dataframe containing the length index from the}
\author{Bjarki Þór Elvarsson}
\arguments{\item{sim}{Results from a Rgadget simulation}
\item{length.groups}{A vector of cutoff points for the length groups}
\item{sigma}{sigma for a log-normal error}}
