% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{toDataFrame}
\alias{toDataFrame}
\title{toDataFrame}
\usage{
toDataFrame(sim)
}
\arguments{
\item{sim}{gadget simulation object (from gadget.simulate)}
}
\value{
data.frame
}
\description{
toDataFrame
}
\author{
Bjarki Thor Elvarsson
}
