% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunc.R
\name{vonB}
\alias{vonB}
\title{Von Bertalanffy}
\usage{
vonB(lsup, k, a)
}
\arguments{
\item{lsup}{\eqn{L_\infty }{L_infty} terminal length}

\item{k}{\eqn{\kappa}{kappa}, the growth parameter}

\item{a}{age of the individual}
}
\value{
a vector of length(a) with the calculated VB curve at age(s) a.
}
\description{
Calculate the Von Bertanlaffy curve according to the formula
\deqn{L(a) = L_\infty (1-e^{-\kappa a})}{L(a) = L_infty (1-e^{-kappa a})}
}
\author{
Bjarki Þór Elvarsson
}
