% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{write.gadget.area}
\alias{write.gadget.area}
\title{write gadget area file}
\usage{
write.gadget.area(area, file = "area")
}
\value{
nothing
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}
