% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetfile.R
\name{write.gadget.file}
\alias{write.gadget.file}
\title{Write gadgetfile to disk, including any dependant files, and update the mainfile}
\usage{
write.gadget.file(obj, path, recursive = TRUE)
}
\arguments{
\item{obj}{gadgetfile object to write}

\item{path}{Base directory to write out to}

\item{recursive}{Write out all nested files too (default TRUE)?}
}
\description{
Write gadgetfile to disk, including any dependant files, and update the mainfile
}
