% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{write.gadget.fleet}
\alias{write.gadget.fleet}
\title{write gadget fleet}
\usage{
write.gadget.fleet(fleet, file = "fleet")
}
\value{
nothing
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Þór Elvarsson
}

