% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.gadget.fleet}
\alias{write.gadget.fleet}
\title{write gadget fleet}
\usage{
write.gadget.fleet(fleet, file = "fleet")
}
\arguments{
\item{fleet}{}

\item{file}{}
}
\value{
nothing
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Þór Elvarsson
}

