% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetFileIO.R
\name{write.gadget.time}
\alias{write.gadget.time}
\title{write gadget time file}
\usage{
write.gadget.time(time, file = "time")
}
\arguments{
\item{time}{data.frame with time}

\item{file}{location of the time file}
}
\value{
nothing
}
\description{
<description>
}
\details{
<details>
}
\author{
Bjarki Thor Elvarsson
}
